==> 03-DivisaoDados.ipynb <==
### Substituído pelo 03-rev1, com trechos otimizados

# Importa as bibliotecas
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
import numpy as np

# Faz a leitura dos dados
cx_alimentacao = pd.read_excel('Dados_Conceicao_Caixa_Alimentacao.xlsx')
cx_recirculacao = pd.read_excel('Dados_Conceicao_Caixa_Recirculacao.xlsx')

target = pd.read_excel('T_Conceicao.xlsx')

# Pasta para armazenamento dos dados
pasta = 'ReducaoDimensao/'

# Data a partir da qual coletar os dados (anterior à primeira análise)
# base = target.iloc[1006][0] ## Abril 01
base = target.iloc[0][0] ## Primeiro dado
base = base - np.timedelta64(1, 'h') ## Antes do primeiro, para que pegue todos os dados disponíveis
bkp_base = base
# Data até a qual coletar os dados (última análise)
# fim = target.iloc[1094,0] ## Abril 10
fim = target.iloc[-1,0]
bkp_fim = fim

rougher1 = pd.read_excel('Dados_Conceicao_Rougher1.xlsx')
rougher2 = pd.read_excel('Dados_Conceicao_Rougher2.xlsx')

cleaner1 = pd.read_excel('Dados_Conceicao_Cleaner1.xlsx')
cleaner2 = pd.read_excel('Dados_Conceicao_Cleaner2.xlsx')
recleaner = pd.read_excel('Dados_Conceicao_Recleaner.xlsx')

scavenger1b1 = pd.read_excel('Dados_Conceicao_Scavenger1_B1.xlsx')
scavenger1b2 = pd.read_excel('Dados_Conceicao_Scavenger1_B2.xlsx')
scavenger1b3 = pd.read_excel('Dados_Conceicao_Scavenger1_B3.xlsx')
scavenger2b1 = pd.read_excel('Dados_Conceicao_Scavenger2_B1.xlsx')
scavenger2b2 = pd.read_excel('Dados_Conceicao_Scavenger2_B2.xlsx')
scavenger2b3 = pd.read_excel('Dados_Conceicao_Scavenger2_B3.xlsx')

%store target
%store cx_alimentacao
%store cx_recirculacao
%store cleaner1
%store cleaner2
%store recleaner
%store rougher1
%store rougher2
%store scavenger1b1
%store scavenger1b2
%store scavenger1b3
%store scavenger2b1
%store scavenger2b2
%store scavenger2b3

%store -r target
%store -r cx_alimentacao
%store -r cx_recirculacao
%store -r cleaner1
%store -r cleaner2
%store -r recleaner
%store -r rougher1
%store -r rougher2
%store -r scavenger1b1
%store -r scavenger1b2
%store -r scavenger1b3
%store -r scavenger2b1
%store -r scavenger2b2
%store -r scavenger2b3

# Transforma a primeira linha em cabeçalho
cx_alimentacao.columns = cx_alimentacao.iloc[0]
cx_recirculacao.columns = cx_recirculacao.iloc[0]

# Remove as duas primeiras linhas, que contêm cabeçalhos e nomes de TAGs
cx_alimentacao = cx_alimentacao.drop([0,1])
cx_recirculacao = cx_recirculacao.drop([0,1])

# Refaz a numeração do index
cx_alimentacao.reset_index(drop=True, inplace=True)
cx_recirculacao.reset_index(drop=True, inplace=True)

# Preenche os dados faltantes com zeros
cx_alimentacao = cx_alimentacao.fillna(0)
cx_recirculacao = cx_recirculacao.fillna(0)

# Transforma a primeira linha em cabeçalho
rougher1.columns = rougher1.iloc[0]
rougher2.columns = rougher2.iloc[0]

# Remove as duas primeiras linhas, que contêm cabeçalhos e nomes de TAGs
rougher1 = rougher1.drop([0,1])
rougher2 = rougher2.drop([0,1])

# Refaz a numeração do index
rougher1.reset_index(drop=True, inplace=True)
rougher2.reset_index(drop=True, inplace=True)

# Preenche os dados faltantes com zeros
rougher1 = rougher1.fillna(0)
rougher2 = rougher2.fillna(0)

x.str.strip()
isinstance(x, str)

selecao = rougher2.iloc[:,0] == isinstance(rougher2.iloc[:,0], str)
rougher2[selecao].apply(lambda x: 'no' if isinstance(x, str) else x )

# Transforma a primeira linha em cabeçalho
cleaner1.columns = cleaner1.iloc[0]
cleaner2.columns = cleaner2.iloc[0]
recleaner.columns = recleaner.iloc[0]

# Remove as duas primeiras linhas, que contêm cabeçalhos e nomes de TAGs
cleaner1 = cleaner1.drop([0,1])
cleaner2 = cleaner2.drop([0,1])
recleaner = recleaner.drop([0,1])

# Refaz a numeração do index
cleaner1.reset_index(drop=True, inplace=True)
cleaner2.reset_index(drop=True, inplace=True)
recleaner.reset_index(drop=True, inplace=True)

# Preenche os dados faltantes com zeros
cleaner1 = cleaner1.fillna(0)
cleaner2 = cleaner2.fillna(0)
recleaner = cleaner2.fillna(0)

# Transforma a primeira linha em cabeçalho
scavenger1b1.columns = scavenger1b1.iloc[0]
scavenger1b2.columns = scavenger1b2.iloc[0]
scavenger1b3.columns = scavenger1b3.iloc[0]
scavenger2b1.columns = scavenger2b1.iloc[0]
scavenger2b2.columns = scavenger2b2.iloc[0]
scavenger2b3.columns = scavenger2b3.iloc[0]

# Remove as duas primeiras linhas, que contêm cabeçalhos e nomes de TAGs
scavenger1b1 = scavenger1b1.drop([0,1])
scavenger1b2 = scavenger1b2.drop([0,1])
scavenger1b3 = scavenger1b3.drop([0,1])
scavenger2b1 = scavenger2b1.drop([0,1])
scavenger2b2 = scavenger2b2.drop([0,1])
scavenger2b3 = scavenger2b3.drop([0,1])

# Refaz a numeração do index
scavenger1b1.reset_index(drop=True, inplace=True)
scavenger1b2.reset_index(drop=True, inplace=True)
scavenger1b3.reset_index(drop=True, inplace=True)
scavenger2b1.reset_index(drop=True, inplace=True)
scavenger2b2.reset_index(drop=True, inplace=True)
scavenger2b3.reset_index(drop=True, inplace=True)

# Preenche os dados faltantes com zeros
scavenger1b1 = scavenger1b1.fillna(0)
scavenger1b2 = scavenger1b2.fillna(0)
scavenger1b3 = scavenger1b3.fillna(0)
scavenger2b1 = scavenger2b1.fillna(0)
scavenger2b2 = scavenger2b2.fillna(0)
scavenger2b3 = scavenger2b3.fillna(0)

# Renomeia o cabeçalho de target
target.columns = ['DataHora', 'Ferro', 'Silica']

# Selecao dos target a partir da data inicial até a data final
selecao = target.iloc[:,0] > base
selecao = selecao & (target.iloc[:,0] <= fim)
target = target[selecao]

# Zera o índice
#target.index = [j for j in range(target.iloc[:,0].shape[0])]
target.reset_index(drop=True, inplace=True)

target.to_excel(pasta + 'T_Conceicao.xlsx', 'Planilha1', index=False)

### Separa as variáveis da caixa de alimentação

# Seleciona duas colunas, formando um agrupamento com os dados de um parâmetro
nivel_cx_alimentacao = cx_alimentacao.iloc[:,[0,1]]
densidade_cx_alimentacao = cx_alimentacao.iloc[:,[2,3]]
vazao_vol_cx_alimentacao = cx_alimentacao.iloc[:,[4,5]]
vazao_mas_cx_alimentacao = cx_alimentacao.iloc[:,[6,7]]
vazao_amina_cx_alimentacao = cx_alimentacao.iloc[:,[8,9]]
vazao_amido_cx_alimentacao = cx_alimentacao.iloc[:,[10,11]]

# Retira os dados nulos
selecao = nivel_cx_alimentacao.iloc[:,0] != 0
nivel_cx_alimentacao = nivel_cx_alimentacao[selecao]

selecao = densidade_cx_alimentacao.iloc[:,0] != 0
densidade_cx_alimentacao = densidade_cx_alimentacao[selecao]

selecao = vazao_vol_cx_alimentacao.iloc[:,0] != 0
vazao_vol_cx_alimentacao = vazao_vol_cx_alimentacao[selecao]

selecao = vazao_mas_cx_alimentacao.iloc[:,0] != 0
vazao_mas_cx_alimentacao = vazao_mas_cx_alimentacao[selecao]

selecao = vazao_amina_cx_alimentacao.iloc[:,0] != 0
vazao_amina_cx_alimentacao = vazao_amina_cx_alimentacao[selecao]

selecao = vazao_amido_cx_alimentacao.iloc[:,0] != 0
vazao_amido_cx_alimentacao = vazao_amido_cx_alimentacao[selecao]

# Remove os dados que estão fora do intervalo
selecao = (nivel_cx_alimentacao.iloc[:,0] > base) & (nivel_cx_alimentacao.iloc[:,0] <= fim)
nivel_cx_alimentacao = nivel_cx_alimentacao[selecao]

selecao = (densidade_cx_alimentacao.iloc[:,0] > base) & (densidade_cx_alimentacao.iloc[:,0] <= fim)
densidade_cx_alimentacao = densidade_cx_alimentacao[selecao]

selecao = (vazao_vol_cx_alimentacao.iloc[:,0] > base) & (vazao_vol_cx_alimentacao.iloc[:,0] <= fim)
vazao_vol_cx_alimentacao = vazao_vol_cx_alimentacao[selecao]

selecao = (vazao_mas_cx_alimentacao.iloc[:,0] > base) & (vazao_mas_cx_alimentacao.iloc[:,0] <= fim)
vazao_mas_cx_alimentacao = vazao_mas_cx_alimentacao[selecao]

selecao = (vazao_amina_cx_alimentacao.iloc[:,0] > base) & (vazao_amina_cx_alimentacao.iloc[:,0] <= fim)
vazao_amina_cx_alimentacao = vazao_amina_cx_alimentacao[selecao]

selecao = (vazao_amido_cx_alimentacao.iloc[:,0] > base) & (vazao_amido_cx_alimentacao.iloc[:,0] <= fim)
vazao_amido_cx_alimentacao = vazao_amido_cx_alimentacao[selecao]

# Zera o índice
#nivel_cx_alimentacao.index = [j for j in range(nivel_cx_alimentacao.iloc[:,0].shape[0])]
#densidade_cx_alimentacao.index = [j for j in range(densidade_cx_alimentacao.iloc[:,0].shape[0])]
#vazao_vol_cx_alimentacao.index = [j for j in range(vazao_vol_cx_alimentacao.iloc[:,0].shape[0])]
#vazao_mas_cx_alimentacao.index = [j for j in range(vazao_mas_cx_alimentacao.iloc[:,0].shape[0])]
#vazao_amina_cx_alimentacao.index = [j for j in range(vazao_amina_cx_alimentacao.iloc[:,0].shape[0])]
#vazao_amido_cx_alimentacao.index = [j for j in range(vazao_amido_cx_alimentacao.iloc[:,0].shape[0])]
nivel_cx_alimentacao.reset_index(drop=True, inplace=True)
densidade_cx_alimentacao.reset_index(drop=True, inplace=True)
vazao_vol_cx_alimentacao.reset_index(drop=True, inplace=True)
vazao_mas_cx_alimentacao.reset_index(drop=True, inplace=True)
vazao_amina_cx_alimentacao.reset_index(drop=True, inplace=True)
vazao_amido_cx_alimentacao.reset_index(drop=True, inplace=True)

# Cria coluna para separação em grupos
nivel_cx_alimentacao['gr'] = -1
densidade_cx_alimentacao['gr'] = -1
vazao_vol_cx_alimentacao['gr'] = -1
vazao_mas_cx_alimentacao['gr'] = -1
vazao_amina_cx_alimentacao['gr'] = -1
vazao_amido_cx_alimentacao['gr'] = -1

i = 0
base = bkp_base
for amostra in target.iloc[:,0]:
    # Critério de seleção dos dados com base na data e hora
    selecao = (nivel_cx_alimentacao.iloc[:,0] > base) & (nivel_cx_alimentacao.iloc[:,0] <= amostra)
    
    nivel_cx_alimentacao.loc[selecao, 'gr'] = i
    
    # Critério de seleção dos dados com base na data e hora
    selecao = (densidade_cx_alimentacao.iloc[:,0] > base) & (densidade_cx_alimentacao.iloc[:,0] <= amostra)
    
    densidade_cx_alimentacao.loc[selecao, 'gr'] = i
    
    # Critério de seleção dos dados com base na data e hora
    selecao = (vazao_vol_cx_alimentacao.iloc[:,0] > base) & (vazao_vol_cx_alimentacao.iloc[:,0] <= amostra)
    
    vazao_vol_cx_alimentacao.loc[selecao, 'gr'] = i
    
    # Critério de seleção dos dados com base na data e hora
    selecao = (vazao_mas_cx_alimentacao.iloc[:,0] > base) & (vazao_mas_cx_alimentacao.iloc[:,0] <= amostra)
    
    vazao_mas_cx_alimentacao.loc[selecao, 'gr'] = i
    
    # Critério de seleção dos dados com base na data e hora
    selecao = (vazao_amina_cx_alimentacao.iloc[:,0] > base) & (vazao_amina_cx_alimentacao.iloc[:,0] <= amostra)
    
    vazao_amina_cx_alimentacao.loc[selecao, 'gr'] = i
    
    # Critério de seleção dos dados com base na data e hora
    selecao = (vazao_amido_cx_alimentacao.iloc[:,0] > base) & (vazao_amido_cx_alimentacao.iloc[:,0] <= amostra)
    
    vazao_amido_cx_alimentacao.loc[selecao, 'gr'] = i
    
    base = amostra
    if i % 5 == 0:
        print(i)
    i += 1

nivel_cx_alimentacao.to_excel(pasta + 'CA_Nivel.xlsx', 'Planilha1', index=False)

densidade_cx_alimentacao.to_excel(pasta + 'CA_Densidade.xlsx', 'Planilha1', index=False)

vazao_vol_cx_alimentacao.to_excel(pasta + 'CA_Vazao_Vol.xlsx', 'Planilha1', index=False)

vazao_mas_cx_alimentacao.to_excel(pasta + 'CA_Vazao_Mas.xlsx', 'Planilha1', index=False)

vazao_amina_cx_alimentacao.to_excel(pasta + 'CA_Vazao_Amina.xlsx', 'Planilha1', index=False)

vazao_amido_cx_alimentacao.to_excel(pasta + 'CA_Vazao_Amido.xlsx', 'Planilha1', index=False)

### Separa as variáveis da caixa de recirculação

base = bkp_base
fim = bkp_fim

# Seleciona duas colunas, formando um agrupamento com os dados de um parâmetro
nivel_cx_recirculacao = cx_recirculacao.iloc[:,[0,1]]
densidade_cx_recirculacao = cx_recirculacao.iloc[:,[2,3]]
vazao_vol_cx_recirculacao = cx_recirculacao.iloc[:,[4,5]]
vazao_mas_cx_recirculacao = cx_recirculacao.iloc[:,[6,7]]

# Retira os dados nulos
selecao = nivel_cx_recirculacao.iloc[:,0] != 0
nivel_cx_recirculacao = nivel_cx_recirculacao[selecao]

selecao = densidade_cx_recirculacao.iloc[:,0] != 0
densidade_cx_recirculacao = densidade_cx_recirculacao[selecao]

selecao = vazao_vol_cx_recirculacao.iloc[:,0] != 0
vazao_vol_cx_recirculacao = vazao_vol_cx_recirculacao[selecao]

selecao = vazao_mas_cx_recirculacao.iloc[:,0] != 0
vazao_mas_cx_recirculacao = vazao_mas_cx_recirculacao[selecao]

# Remove os dados que estão fora do intervalo
selecao = (nivel_cx_recirculacao.iloc[:,0] > base) & (nivel_cx_recirculacao.iloc[:,0] <= fim)
nivel_cx_recirculacao = nivel_cx_recirculacao[selecao]

selecao = (densidade_cx_recirculacao.iloc[:,0] > base) & (densidade_cx_recirculacao.iloc[:,0] <= fim)
densidade_cx_recirculacao = densidade_cx_recirculacao[selecao]

selecao = (vazao_vol_cx_recirculacao.iloc[:,0] > base) & (vazao_vol_cx_recirculacao.iloc[:,0] <= fim)
vazao_vol_cx_recirculacao = vazao_vol_cx_recirculacao[selecao]

selecao = (vazao_mas_cx_recirculacao.iloc[:,0] > base) & (vazao_mas_cx_recirculacao.iloc[:,0] <= fim)
vazao_mas_cx_recirculacao = vazao_mas_cx_recirculacao[selecao]

# Zera o índice
#nivel_cx_recirculacao.index = [j for j in range(nivel_cx_recirculacao.iloc[:,0].shape[0])]
#densidade_cx_recirculacao.index = [j for j in range(densidade_cx_recirculacao.iloc[:,0].shape[0])]
#vazao_vol_cx_recirculacao.index = [j for j in range(vazao_vol_cx_recirculacao.iloc[:,0].shape[0])]
#vazao_mas_cx_recirculacao.index = [j for j in range(vazao_mas_cx_recirculacao.iloc[:,0].shape[0])]
nivel_cx_recirculacao.reset_index(drop=True, inplace=True)
densidade_cx_recirculacao.reset_index(drop=True, inplace=True)
vazao_vol_cx_recirculacao.reset_index(drop=True, inplace=True)
vazao_mas_cx_recirculacao.reset_index(drop=True, inplace=True)

# Cria coluna para separação em grupos
nivel_cx_recirculacao['gr'] = -1
densidade_cx_recirculacao['gr'] = -1
vazao_vol_cx_recirculacao['gr'] = -1
vazao_mas_cx_recirculacao['gr'] = -1

i = 0
base = bkp_base
for amostra in target.iloc[:,0]:
    # Critério de seleção dos dados com base na data e hora
    selecao = (nivel_cx_recirculacao.iloc[:,0] > base) & (nivel_cx_recirculacao.iloc[:,0] <= amostra)
    
    nivel_cx_recirculacao.loc[selecao, 'gr'] = i
    
    # Critério de seleção dos dados com base na data e hora
    selecao = (densidade_cx_recirculacao.iloc[:,0] > base) & (densidade_cx_recirculacao.iloc[:,0] <= amostra)
    
    densidade_cx_recirculacao.loc[selecao, 'gr'] = i
    
    # Critério de seleção dos dados com base na data e hora
    selecao = (vazao_vol_cx_recirculacao.iloc[:,0] > base) & (vazao_vol_cx_recirculacao.iloc[:,0] <= amostra)
    
    vazao_vol_cx_recirculacao.loc[selecao, 'gr'] = i
    
    # Critério de seleção dos dados com base na data e hora
    selecao = (vazao_mas_cx_recirculacao.iloc[:,0] > base) & (vazao_mas_cx_recirculacao.iloc[:,0] <= amostra)
    
    vazao_mas_cx_recirculacao.loc[selecao, 'gr'] = i
    
    base = amostra
    if i % 5 == 0:
        print(i)
    i += 1

nivel_cx_recirculacao.to_excel(pasta + 'CR_Nivel.xlsx', 'Planilha1', index=False)

densidade_cx_recirculacao.to_excel(pasta + 'CR_Densidade.xlsx', 'Planilha1', index=False)

vazao_vol_cx_recirculacao.to_excel(pasta + 'CR_Vazao_Vol.xlsx', 'Planilha1', index=False)

vazao_mas_cx_recirculacao.to_excel(pasta + 'CR_Vazao_Mas.xlsx', 'Planilha1', index=False)

### Separa as variáveis dos Roughers

base = bkp_base
fim = bkp_fim

# Seleciona duas colunas, formando um agrupamento com os dados de um parâmetro
nivel_espuma_rougher1 = rougher1.iloc[:,[0,1]]
vazao_ar_rougher1 = rougher1.iloc[:,[2,3]]
valvula_dardo1_rougher1 = rougher1.iloc[:,[4,5]]
valvula_dardo2_rougher1 = rougher1.iloc[:,[6,7]]
valvula_ar_rougher1 = rougher1.iloc[:,[8,9]]

# Retira os dados nulos
selecao = (nivel_espuma_rougher1.iloc[:,0] != 0) & (nivel_espuma_rougher1.iloc[:,0] != " ")
nivel_espuma_rougher1 = nivel_espuma_rougher1[selecao]

selecao = (vazao_ar_rougher1.iloc[:,0] != 0) & (vazao_ar_rougher1.iloc[:,0] != " ")
vazao_ar_rougher1 = vazao_ar_rougher1[selecao]

selecao = (valvula_dardo1_rougher1.iloc[:,0] != 0) & (valvula_dardo1_rougher1.iloc[:,0] != " ")
valvula_dardo1_rougher1 = valvula_dardo1_rougher1[selecao]

selecao = (valvula_dardo2_rougher1.iloc[:,0] != 0) & (valvula_dardo2_rougher1.iloc[:,0] != " ")
valvula_dardo2_rougher1 = valvula_dardo2_rougher1[selecao]

selecao = (valvula_ar_rougher1.iloc[:,0] != 0) & (valvula_ar_rougher1.iloc[:,0] != " ")
valvula_ar_rougher1 = valvula_ar_rougher1[selecao]

# Seleciona duas colunas, formando um agrupamento com os dados de um parâmetro
nivel_espuma_rougher2 = rougher2.iloc[:,[0,1]]
vazao_ar_rougher2 = rougher2.iloc[:,[2,3]]
valvula_dardo1_rougher2 = rougher2.iloc[:,[4,5]]
valvula_dardo2_rougher2 = rougher2.iloc[:,[6,7]]
valvula_ar_rougher2 = rougher2.iloc[:,[8,9]]

# Retira os dados nulos
selecao = (nivel_espuma_rougher2.iloc[:,0] != 0) & (nivel_espuma_rougher2.iloc[:,0] != " ")
nivel_espuma_rougher2 = nivel_espuma_rougher2[selecao]

selecao = (vazao_ar_rougher2.iloc[:,0] != 0) & (vazao_ar_rougher2.iloc[:,0] != " ")
vazao_ar_rougher2 = vazao_ar_rougher2[selecao]

selecao = (valvula_dardo1_rougher2.iloc[:,0] != 0) & (valvula_dardo1_rougher2.iloc[:,0] != " ")
valvula_dardo1_rougher2 = valvula_dardo1_rougher2[selecao]

selecao = (valvula_dardo2_rougher2.iloc[:,0] != 0) & (valvula_dardo2_rougher2.iloc[:,0] != " ")
valvula_dardo2_rougher2 = valvula_dardo2_rougher2[selecao]

selecao = (valvula_ar_rougher2.iloc[:,0] != 0) & (valvula_ar_rougher2.iloc[:,0] != " ")
valvula_ar_rougher2 = valvula_ar_rougher2[selecao]

# Remove os dados que estão fora do intervalo
selecao = (nivel_espuma_rougher1.iloc[:,0] > base) & (nivel_espuma_rougher1.iloc[:,0] <= fim)
nivel_espuma_rougher1 = nivel_espuma_rougher1[selecao]

selecao = (vazao_ar_rougher1.iloc[:,0] > base) & (vazao_ar_rougher1.iloc[:,0] <= fim)
vazao_ar_rougher1 = vazao_ar_rougher1[selecao]

selecao = (valvula_dardo1_rougher1.iloc[:,0] > base) & (valvula_dardo1_rougher1.iloc[:,0] <= fim)
valvula_dardo1_rougher1 = valvula_dardo1_rougher1[selecao]

selecao = (valvula_dardo2_rougher1.iloc[:,0] > base) & (valvula_dardo2_rougher1.iloc[:,0] <= fim)
valvula_dardo2_rougher1 = valvula_dardo2_rougher1[selecao]

selecao = (valvula_ar_rougher1.iloc[:,0] > base) & (valvula_ar_rougher1.iloc[:,0] <= fim)
valvula_ar_rougher1 = valvula_ar_rougher1[selecao]

# Remove os dados que estão fora do intervalo
selecao = (nivel_espuma_rougher2.iloc[:,0] > base) & (nivel_espuma_rougher2.iloc[:,0] <= fim)
nivel_espuma_rougher2 = nivel_espuma_rougher2[selecao]

selecao = (vazao_ar_rougher2.iloc[:,0] > base) & (vazao_ar_rougher2.iloc[:,0] <= fim)
vazao_ar_rougher2 = vazao_ar_rougher2[selecao]

selecao = (valvula_dardo1_rougher2.iloc[:,0] > base) & (valvula_dardo1_rougher2.iloc[:,0] <= fim)
valvula_dardo1_rougher2 = valvula_dardo1_rougher2[selecao]

selecao = (valvula_dardo2_rougher2.iloc[:,0] > base) & (valvula_dardo2_rougher2.iloc[:,0] <= fim)
valvula_dardo2_rougher2 = valvula_dardo2_rougher2[selecao]

selecao = (valvula_ar_rougher2.iloc[:,0] > base) & (valvula_ar_rougher2.iloc[:,0] <= fim)
valvula_ar_rougher2 = valvula_ar_rougher2[selecao]

# Zera o índice
#nivel_espuma_rougher1.index = [j for j in range(nivel_espuma_rougher1.iloc[:,0].shape[0])]
#vazao_ar_rougher1.index = [j for j in range(vazao_ar_rougher1.iloc[:,0].shape[0])]
#valvula_dardo1_rougher1.index = [j for j in range(valvula_dardo1_rougher1.iloc[:,0].shape[0])]
#valvula_dardo2_rougher1.index = [j for j in range(valvula_dardo2_rougher1.iloc[:,0].shape[0])]
#valvula_ar_rougher1.index = [j for j in range(valvula_ar_rougher1.iloc[:,0].shape[0])]
nivel_espuma_rougher1.reset_index(drop=True, inplace=True)
vazao_ar_rougher1.reset_index(drop=True, inplace=True)
valvula_dardo1_rougher1.reset_index(drop=True, inplace=True)
valvula_dardo2_rougher1.reset_index(drop=True, inplace=True)
valvula_ar_rougher1.reset_index(drop=True, inplace=True)

# Zera o índice
#nivel_espuma_rougher2.index = [j for j in range(nivel_espuma_rougher2.iloc[:,0].shape[0])]
#vazao_ar_rougher2.index = [j for j in range(vazao_ar_rougher2.iloc[:,0].shape[0])]
#valvula_dardo1_rougher2.index = [j for j in range(valvula_dardo1_rougher2.iloc[:,0].shape[0])]
#valvula_dardo2_rougher2.index = [j for j in range(valvula_dardo2_rougher2.iloc[:,0].shape[0])]
#valvula_ar_rougher2.index = [j for j in range(valvula_ar_rougher2.iloc[:,0].shape[0])]
nivel_espuma_rougher2.reset_index(drop=True, inplace=True)
vazao_ar_rougher2.reset_index(drop=True, inplace=True)
valvula_dardo1_rougher2.reset_index(drop=True, inplace=True)
valvula_dardo2_rougher2.reset_index(drop=True, inplace=True)
valvula_ar_rougher2.reset_index(drop=True, inplace=True)

# Cria coluna para separação em grupos
nivel_espuma_rougher1['gr'] = -1
vazao_ar_rougher1['gr'] = -1
valvula_dardo1_rougher1['gr'] = -1
valvula_dardo2_rougher1['gr'] = -1
valvula_ar_rougher1['gr'] = -1

# Cria coluna para separação em grupos
nivel_espuma_rougher2['gr'] = -1
vazao_ar_rougher2['gr'] = -1
valvula_dardo1_rougher2['gr'] = -1
valvula_dardo2_rougher2['gr'] = -1
valvula_ar_rougher2['gr'] = -1

i = 0
base = bkp_base
for amostra in target.iloc[:,0]:
    # Critério de seleção dos dados com base na data e hora
    selecao = (nivel_espuma_rougher1.iloc[:,0] > base) & (nivel_espuma_rougher1.iloc[:,0] <= amostra)
    
    nivel_espuma_rougher1.loc[selecao, 'gr'] = i
    
    # Critério de seleção dos dados com base na data e hora
    selecao = (vazao_ar_rougher1.iloc[:,0] > base) & (vazao_ar_rougher1.iloc[:,0] <= amostra)
    
    vazao_ar_rougher1.loc[selecao, 'gr'] = i
    
    # Critério de seleção dos dados com base na data e hora
    selecao = (valvula_dardo1_rougher1.iloc[:,0] > base) & (valvula_dardo1_rougher1.iloc[:,0] <= amostra)
    
    valvula_dardo1_rougher1.loc[selecao, 'gr'] = i
    
    # Critério de seleção dos dados com base na data e hora
    selecao = (valvula_dardo2_rougher1.iloc[:,0] > base) & (valvula_dardo2_rougher1.iloc[:,0] <= amostra)
    
    valvula_dardo2_rougher1.loc[selecao, 'gr'] = i
    
    # Critério de seleção dos dados com base na data e hora
    selecao = (valvula_ar_rougher1.iloc[:,0] > base) & (valvula_ar_rougher1.iloc[:,0] <= amostra)
    
    valvula_ar_rougher1.loc[selecao, 'gr'] = i
    
    base = amostra
    if i % 5 == 0:
        print(i)
    i += 1

i = 0
base = bkp_base
for amostra in target.iloc[:,0]:
    # Critério de seleção dos dados com base na data e hora
    selecao = (nivel_espuma_rougher2.iloc[:,0] > base) & (nivel_espuma_rougher2.iloc[:,0] <= amostra)
    
    nivel_espuma_rougher2.loc[selecao, 'gr'] = i
    
    # Critério de seleção dos dados com base na data e hora
    selecao = (vazao_ar_rougher2.iloc[:,0] > base) & (vazao_ar_rougher2.iloc[:,0] <= amostra)
    
    vazao_ar_rougher2.loc[selecao, 'gr'] = i
    
    # Critério de seleção dos dados com base na data e hora
    selecao = (valvula_dardo1_rougher2.iloc[:,0] > base) & (valvula_dardo1_rougher2.iloc[:,0] <= amostra)
    
    valvula_dardo1_rougher2.loc[selecao, 'gr'] = i
    
    # Critério de seleção dos dados com base na data e hora
    selecao = (valvula_dardo2_rougher2.iloc[:,0] > base) & (valvula_dardo2_rougher2.iloc[:,0] <= amostra)
    
    valvula_dardo2_rougher2.loc[selecao, 'gr'] = i
    
    # Critério de seleção dos dados com base na data e hora
    selecao = (valvula_ar_rougher2.iloc[:,0] > base) & (valvula_ar_rougher2.iloc[:,0] <= amostra)
    
    valvula_ar_rougher2.loc[selecao, 'gr'] = i
    
    base = amostra
    if i % 5 == 0:
        print(i)
    i += 1

nivel_espuma_rougher1.to_excel(pasta + 'R1_Nivel_Espuma.xlsx', 'Planilha1', index=False)

vazao_ar_rougher1.to_excel(pasta + 'R1_Vazao_Ar.xlsx', 'Planilha1', index=False)

valvula_dardo1_rougher1.to_excel(pasta + 'R1_Valv_Dardo1.xlsx', 'Planilha1', index=False)

valvula_dardo2_rougher1.to_excel(pasta + 'R1_Valv_Dardo2.xlsx', 'Planilha1', index=False)

valvula_ar_rougher1.to_excel(pasta + 'R1_Valv_Ar.xlsx', 'Planilha1', index=False)

nivel_espuma_rougher2.to_excel(pasta + 'R2_Nivel_Espuma.xlsx', 'Planilha1', index=False)

vazao_ar_rougher2.to_excel(pasta + 'R2_Vazao_Ar.xlsx', 'Planilha1', index=False)

valvula_dardo1_rougher2.to_excel(pasta + 'R2_Valv_Dardo1.xlsx', 'Planilha1', index=False)

valvula_dardo2_rougher2.to_excel(pasta + 'R2_Valv_Dardo2.xlsx', 'Planilha1', index=False)

valvula_ar_rougher2.to_excel(pasta + 'R2_Valv_Ar.xlsx', 'Planilha1', index=False)

### Separa as variáveis dos cleaners e recleaner

base = bkp_base
fim = bkp_fim

# Seleciona duas colunas, formando um agrupamento com os dados de um parâmetro
nivel_espuma_cleaner1 = cleaner1.iloc[:,[0,1]]
vazao_ar_cleaner1 = cleaner1.iloc[:,[2,3]]
valvula_dardo1_cleaner1 = cleaner1.iloc[:,[4,5]]
valvula_dardo2_cleaner1 = cleaner1.iloc[:,[6,7]]
valvula_ar_cleaner1 = cleaner1.iloc[:,[8,9]]

# Retira os dados nulos
selecao = (nivel_espuma_cleaner1.iloc[:,0] != 0) & (nivel_espuma_cleaner1.iloc[:,0] != " ")
nivel_espuma_cleaner1 = nivel_espuma_cleaner1[selecao]

selecao = (vazao_ar_cleaner1.iloc[:,0] != 0) & (vazao_ar_cleaner1.iloc[:,0] != " ")
vazao_ar_cleaner1 = vazao_ar_cleaner1[selecao]

selecao = (valvula_dardo1_cleaner1.iloc[:,0] != 0) & (valvula_dardo1_cleaner1.iloc[:,0] != " ")
valvula_dardo1_cleaner1 = valvula_dardo1_cleaner1[selecao]

selecao = (valvula_dardo2_cleaner1.iloc[:,0] != 0) & (valvula_dardo2_cleaner1.iloc[:,0] != " ")
valvula_dardo2_cleaner1 = valvula_dardo2_cleaner1[selecao]

selecao = (valvula_ar_cleaner1.iloc[:,0] != 0) & (valvula_ar_cleaner1.iloc[:,0] != " ")
valvula_ar_cleaner1 = valvula_ar_cleaner1[selecao]

# Seleciona duas colunas, formando um agrupamento com os dados de um parâmetro
nivel_espuma_cleaner2 = cleaner2.iloc[:,[0,1]]
vazao_ar_cleaner2 = cleaner2.iloc[:,[2,3]]
valvula_dardo1_cleaner2 = cleaner2.iloc[:,[4,5]]
valvula_dardo2_cleaner2 = cleaner2.iloc[:,[6,7]]
valvula_ar_cleaner2 = cleaner2.iloc[:,[8,9]]

# Retira os dados nulos
selecao = (nivel_espuma_cleaner2.iloc[:,0] != 0) & (nivel_espuma_cleaner2.iloc[:,0] != " ")
nivel_espuma_cleaner2 = nivel_espuma_cleaner2[selecao]

selecao = (vazao_ar_cleaner2.iloc[:,0] != 0) & (vazao_ar_cleaner2.iloc[:,0] != " ")
vazao_ar_cleaner2 = vazao_ar_cleaner2[selecao]

selecao = (valvula_dardo1_cleaner2.iloc[:,0] != 0) & (valvula_dardo1_cleaner2.iloc[:,0] != " ")
valvula_dardo1_cleaner2 = valvula_dardo1_cleaner2[selecao]

selecao = (valvula_dardo2_cleaner2.iloc[:,0] != 0) & (valvula_dardo2_cleaner2.iloc[:,0] != " ")
valvula_dardo2_cleaner2 = valvula_dardo2_cleaner2[selecao]

selecao = (valvula_ar_cleaner2.iloc[:,0] != 0) & (valvula_ar_cleaner2.iloc[:,0] != " ")
valvula_ar_cleaner2 = valvula_ar_cleaner2[selecao]

# Seleciona duas colunas, formando um agrupamento com os dados de um parâmetro
nivel_espuma_recleaner = recleaner.iloc[:,[0,1]]
vazao_ar_recleaner = recleaner.iloc[:,[2,3]]
valvula_dardo1_recleaner = recleaner.iloc[:,[4,5]]
valvula_dardo2_recleaner = recleaner.iloc[:,[6,7]]
valvula_ar_recleaner = recleaner.iloc[:,[8,9]]

# Retira os dados nulos
selecao = (nivel_espuma_recleaner.iloc[:,0] != 0) & (nivel_espuma_recleaner.iloc[:,0] != " ")
nivel_espuma_recleaner = nivel_espuma_recleaner[selecao]

selecao = (vazao_ar_recleaner.iloc[:,0] != 0) & (vazao_ar_recleaner.iloc[:,0] != " ")
vazao_ar_recleaner = vazao_ar_recleaner[selecao]

selecao = (valvula_dardo1_recleaner.iloc[:,0] != 0) & (valvula_dardo1_recleaner.iloc[:,0] != " ")
valvula_dardo1_recleaner = valvula_dardo1_recleaner[selecao]

selecao = (valvula_dardo2_recleaner.iloc[:,0] != 0) & (valvula_dardo2_recleaner.iloc[:,0] != " ")
valvula_dardo2_recleaner = valvula_dardo2_recleaner[selecao]

selecao = (valvula_ar_recleaner.iloc[:,0] != 0) & (valvula_ar_recleaner.iloc[:,0] != " ")
valvula_ar_recleaner = valvula_ar_recleaner[selecao]

# Remove os dados que estão fora do intervalo
selecao = (nivel_espuma_cleaner1.iloc[:,0] > base) & (nivel_espuma_cleaner1.iloc[:,0] <= fim)
nivel_espuma_cleaner1 = nivel_espuma_cleaner1[selecao]

selecao = (vazao_ar_cleaner1.iloc[:,0] > base) & (vazao_ar_cleaner1.iloc[:,0] <= fim)
vazao_ar_cleaner1 = vazao_ar_cleaner1[selecao]

selecao = (valvula_dardo1_cleaner1.iloc[:,0] > base) & (valvula_dardo1_cleaner1.iloc[:,0] <= fim)
valvula_dardo1_cleaner1 = valvula_dardo1_cleaner1[selecao]

selecao = (valvula_dardo2_cleaner1.iloc[:,0] > base) & (valvula_dardo2_cleaner1.iloc[:,0] <= fim)
valvula_dardo2_cleaner1 = valvula_dardo2_cleaner1[selecao]

selecao = (valvula_ar_cleaner1.iloc[:,0] > base) & (valvula_ar_cleaner1.iloc[:,0] <= fim)
valvula_ar_cleaner1 = valvula_ar_cleaner1[selecao]

# Remove os dados que estão fora do intervalo
selecao = (nivel_espuma_cleaner2.iloc[:,0] > base) & (nivel_espuma_cleaner2.iloc[:,0] <= fim)
nivel_espuma_cleaner2 = nivel_espuma_cleaner2[selecao]

selecao = (vazao_ar_cleaner2.iloc[:,0] > base) & (vazao_ar_cleaner2.iloc[:,0] <= fim)
vazao_ar_cleaner2 = vazao_ar_cleaner2[selecao]

selecao = (valvula_dardo1_cleaner2.iloc[:,0] > base) & (valvula_dardo1_cleaner2.iloc[:,0] <= fim)
valvula_dardo1_cleaner2 = valvula_dardo1_cleaner2[selecao]

selecao = (valvula_dardo2_cleaner2.iloc[:,0] > base) & (valvula_dardo2_cleaner2.iloc[:,0] <= fim)
valvula_dardo2_cleaner2 = valvula_dardo2_cleaner2[selecao]

selecao = (valvula_ar_cleaner2.iloc[:,0] > base) & (valvula_ar_cleaner2.iloc[:,0] <= fim)
valvula_ar_cleaner2 = valvula_ar_cleaner2[selecao]

# Remove os dados que estão fora do intervalo
selecao = (nivel_espuma_recleaner.iloc[:,0] > base) & (nivel_espuma_recleaner.iloc[:,0] <= fim)
nivel_espuma_recleaner = nivel_espuma_recleaner[selecao]

selecao = (vazao_ar_recleaner.iloc[:,0] > base) & (vazao_ar_recleaner.iloc[:,0] <= fim)
vazao_ar_recleaner = vazao_ar_recleaner[selecao]

selecao = (valvula_dardo1_recleaner.iloc[:,0] > base) & (valvula_dardo1_recleaner.iloc[:,0] <= fim)
valvula_dardo1_recleaner = valvula_dardo1_recleaner[selecao]

selecao = (valvula_dardo2_recleaner.iloc[:,0] > base) & (valvula_dardo2_recleaner.iloc[:,0] <= fim)
valvula_dardo2_recleaner = valvula_dardo2_recleaner[selecao]

selecao = (valvula_ar_recleaner.iloc[:,0] > base) & (valvula_ar_recleaner.iloc[:,0] <= fim)
valvula_ar_recleaner = valvula_ar_recleaner[selecao]

# Zera o índice
#nivel_espuma_cleaner1.index = [j for j in range(nivel_espuma_cleaner1.iloc[:,0].shape[0])]
#vazao_ar_cleaner1.index = [j for j in range(vazao_ar_cleaner1.iloc[:,0].shape[0])]
#valvula_dardo1_cleaner1.index = [j for j in range(valvula_dardo1_cleaner1.iloc[:,0].shape[0])]
#valvula_dardo2_cleaner1.index = [j for j in range(valvula_dardo2_cleaner1.iloc[:,0].shape[0])]
#valvula_ar_cleaner1.index = [j for j in range(valvula_ar_cleaner1.iloc[:,0].shape[0])]
nivel_espuma_cleaner1.reset_index(drop=True, inplace=True)
vazao_ar_cleaner1.reset_index(drop=True, inplace=True)
valvula_dardo1_cleaner1.reset_index(drop=True, inplace=True)
valvula_dardo2_cleaner1.reset_index(drop=True, inplace=True)
valvula_ar_cleaner1.reset_index(drop=True, inplace=True)

# Zera o índice
#nivel_espuma_cleaner2.index = [j for j in range(nivel_espuma_cleaner2.iloc[:,0].shape[0])]
#vazao_ar_cleaner2.index = [j for j in range(vazao_ar_cleaner2.iloc[:,0].shape[0])]
#valvula_dardo1_cleaner2.index = [j for j in range(valvula_dardo1_cleaner2.iloc[:,0].shape[0])]
#valvula_dardo2_cleaner2.index = [j for j in range(valvula_dardo2_cleaner2.iloc[:,0].shape[0])]
#valvula_ar_cleaner2.index = [j for j in range(valvula_ar_cleaner2.iloc[:,0].shape[0])]
nivel_espuma_cleaner2.reset_index(drop=True, inplace=True)
vazao_ar_cleaner2.reset_index(drop=True, inplace=True)
valvula_dardo1_cleaner2.reset_index(drop=True, inplace=True)
valvula_dardo2_cleaner2.reset_index(drop=True, inplace=True)
valvula_ar_cleaner2.reset_index(drop=True, inplace=True)

# Zera o índice
#nivel_espuma_recleaner.index = [j for j in range(nivel_espuma_recleaner.iloc[:,0].shape[0])]
#vazao_ar_recleaner.index = [j for j in range(vazao_ar_recleaner.iloc[:,0].shape[0])]
#valvula_dardo1_recleaner.index = [j for j in range(valvula_dardo1_recleaner.iloc[:,0].shape[0])]
#valvula_dardo2_recleaner.index = [j for j in range(valvula_dardo2_recleaner.iloc[:,0].shape[0])]
#valvula_ar_recleaner.index = [j for j in range(valvula_ar_recleaner.iloc[:,0].shape[0])]
nivel_espuma_recleaner.reset_index(drop=True, inplace=True)
vazao_ar_recleaner.reset_index(drop=True, inplace=True)
valvula_dardo1_recleaner.reset_index(drop=True, inplace=True)
valvula_dardo2_recleaner.reset_index(drop=True, inplace=True)
valvula_ar_recleaner.reset_index(drop=True, inplace=True)

# Cria coluna para separação em grupos
nivel_espuma_cleaner1['gr'] = -1
vazao_ar_cleaner1['gr'] = -1
valvula_dardo1_cleaner1['gr'] = -1
valvula_dardo2_cleaner1['gr'] = -1
valvula_ar_cleaner1['gr'] = -1

# Cria coluna para separação em grupos
nivel_espuma_cleaner2['gr'] = -1
vazao_ar_cleaner2['gr'] = -1
valvula_dardo1_cleaner2['gr'] = -1
valvula_dardo2_cleaner2['gr'] = -1
valvula_ar_cleaner2['gr'] = -1

# Cria coluna para separação em grupos
nivel_espuma_recleaner['gr'] = -1
vazao_ar_recleaner['gr'] = -1
valvula_dardo1_recleaner['gr'] = -1
valvula_dardo2_recleaner['gr'] = -1
valvula_ar_recleaner['gr'] = -1

i = 0
base = bkp_base
for amostra in target.iloc[:,0]:
    # Critério de seleção dos dados com base na data e hora
    selecao = (nivel_espuma_cleaner1.iloc[:,0] > base) & (nivel_espuma_cleaner1.iloc[:,0] <= amostra)
    
    nivel_espuma_cleaner1.loc[selecao, 'gr'] = i
    
    # Critério de seleção dos dados com base na data e hora
    selecao = (vazao_ar_cleaner1.iloc[:,0] > base) & (vazao_ar_cleaner1.iloc[:,0] <= amostra)
    
    vazao_ar_cleaner1.loc[selecao, 'gr'] = i
    
    # Critério de seleção dos dados com base na data e hora
    selecao = (valvula_dardo1_cleaner1.iloc[:,0] > base) & (valvula_dardo1_cleaner1.iloc[:,0] <= amostra)
    
    valvula_dardo1_cleaner1.loc[selecao, 'gr'] = i
    
    # Critério de seleção dos dados com base na data e hora
    selecao = (valvula_dardo2_cleaner1.iloc[:,0] > base) & (valvula_dardo2_cleaner1.iloc[:,0] <= amostra)
    
    valvula_dardo2_cleaner1.loc[selecao, 'gr'] = i
    
    # Critério de seleção dos dados com base na data e hora
    selecao = (valvula_ar_cleaner1.iloc[:,0] > base) & (valvula_ar_cleaner1.iloc[:,0] <= amostra)
    
    valvula_ar_cleaner1.loc[selecao, 'gr'] = i
    
    base = amostra
    if i % 5 == 0:
        print(i)
    i += 1

i = 0
base = bkp_base
for amostra in target.iloc[:,0]:
    # Critério de seleção dos dados com base na data e hora
    selecao = (nivel_espuma_cleaner2.iloc[:,0] > base) & (nivel_espuma_cleaner2.iloc[:,0] <= amostra)
    
    nivel_espuma_cleaner2.loc[selecao, 'gr'] = i
    
    # Critério de seleção dos dados com base na data e hora
    selecao = (vazao_ar_cleaner2.iloc[:,0] > base) & (vazao_ar_cleaner2.iloc[:,0] <= amostra)
    
    vazao_ar_cleaner2.loc[selecao, 'gr'] = i
    
    # Critério de seleção dos dados com base na data e hora
    selecao = (valvula_dardo1_cleaner2.iloc[:,0] > base) & (valvula_dardo1_cleaner2.iloc[:,0] <= amostra)
    
    valvula_dardo1_cleaner2.loc[selecao, 'gr'] = i
    
    # Critério de seleção dos dados com base na data e hora
    selecao = (valvula_dardo2_cleaner2.iloc[:,0] > base) & (valvula_dardo2_cleaner2.iloc[:,0] <= amostra)
    
    valvula_dardo2_cleaner2.loc[selecao, 'gr'] = i
    
    # Critério de seleção dos dados com base na data e hora
    selecao = (valvula_ar_cleaner2.iloc[:,0] > base) & (valvula_ar_cleaner2.iloc[:,0] <= amostra)
    
    valvula_ar_cleaner2.loc[selecao, 'gr'] = i
    
    base = amostra
    if i % 5 == 0:
        print(i)
    i += 1

i = 0
base = bkp_base
for amostra in target.iloc[:,0]:
    # Critério de seleção dos dados com base na data e hora
    selecao = (nivel_espuma_recleaner.iloc[:,0] > base) & (nivel_espuma_recleaner.iloc[:,0] <= amostra)
    
    nivel_espuma_recleaner.loc[selecao, 'gr'] = i
    
    # Critério de seleção dos dados com base na data e hora
    selecao = (vazao_ar_recleaner.iloc[:,0] > base) & (vazao_ar_recleaner.iloc[:,0] <= amostra)
    
    vazao_ar_recleaner.loc[selecao, 'gr'] = i
    
    # Critério de seleção dos dados com base na data e hora
    selecao = (valvula_dardo1_recleaner.iloc[:,0] > base) & (valvula_dardo1_recleaner.iloc[:,0] <= amostra)
    
    valvula_dardo1_recleaner.loc[selecao, 'gr'] = i
    
    # Critério de seleção dos dados com base na data e hora
    selecao = (valvula_dardo2_recleaner.iloc[:,0] > base) & (valvula_dardo2_recleaner.iloc[:,0] <= amostra)
    
    valvula_dardo2_recleaner.loc[selecao, 'gr'] = i
    
    # Critério de seleção dos dados com base na data e hora
    selecao = (valvula_ar_recleaner.iloc[:,0] > base) & (valvula_ar_recleaner.iloc[:,0] <= amostra)
    
    valvula_ar_recleaner.loc[selecao, 'gr'] = i
    
    base = amostra
    if i % 5 == 0:
        print(i)
    i += 1

nivel_espuma_cleaner1.to_excel(pasta + 'C1_Nivel_Espuma.xlsx', 'Planilha1', index=False)

vazao_ar_cleaner1.to_excel(pasta + 'C1_Vazao_Ar.xlsx', 'Planilha1', index=False)

valvula_dardo1_cleaner1.to_excel(pasta + 'C1_Valv_Dardo1.xlsx', 'Planilha1', index=False)

valvula_dardo2_cleaner1.to_excel(pasta + 'C1_Valv_Dardo2.xlsx', 'Planilha1', index=False)

valvula_ar_cleaner1.to_excel(pasta + 'C1_Valv_Ar.xlsx', 'Planilha1', index=False)

nivel_espuma_cleaner2.to_excel(pasta + 'C2_Nivel_Espuma.xlsx', 'Planilha1', index=False)

vazao_ar_cleaner2.to_excel(pasta + 'C2_Vazao_Ar.xlsx', 'Planilha1', index=False)

valvula_dardo1_cleaner2.to_excel(pasta + 'C2_Valv_Dardo1.xlsx', 'Planilha1', index=False)

valvula_dardo2_cleaner2.to_excel(pasta + 'C2_Valv_Dardo2.xlsx', 'Planilha1', index=False)

valvula_ar_cleaner2.to_excel(pasta + 'C2_Valv_Ar.xlsx', 'Planilha1', index=False)

nivel_espuma_recleaner.to_excel(pasta + 'RC_Nivel_Espuma.xlsx', 'Planilha1', index=False)

vazao_ar_recleaner.to_excel(pasta + 'RC_Vazao_Ar.xlsx', 'Planilha1', index=False)

valvula_dardo1_recleaner.to_excel(pasta + 'RC_Valv_Dardo1.xlsx', 'Planilha1', index=False)

valvula_dardo2_recleaner.to_excel(pasta + 'RC_Valv_Dardo2.xlsx', 'Planilha1', index=False)

valvula_ar_recleaner.to_excel(pasta + 'RC_Valv_Ar.xlsx', 'Planilha1', index=False)

### Separa as variáveis dos scavengers

base = bkp_base
fim = bkp_fim

# Seleciona duas colunas, formando um agrupamento com os dados de um parâmetro
nivel_espuma_scavenger1b1 = scavenger1b1.iloc[:,[0,1]]
vazao_ar_scavenger1b1 = scavenger1b1.iloc[:,[2,3]]
valvula_dardo1_scavenger1b1 = scavenger1b1.iloc[:,[4,5]]
valvula_dardo2_scavenger1b1 = scavenger1b1.iloc[:,[6,7]]
valvula_ar_scavenger1b1 = scavenger1b1.iloc[:,[8,9]]

# Retira os dados nulos
selecao = (nivel_espuma_scavenger1b1.iloc[:,0] != 0) & (nivel_espuma_scavenger1b1.iloc[:,0] != " ")
nivel_espuma_scavenger1b1 = nivel_espuma_scavenger1b1[selecao]

selecao = (vazao_ar_scavenger1b1.iloc[:,0] != 0) & (vazao_ar_scavenger1b1.iloc[:,0] != " ")
vazao_ar_scavenger1b1 = vazao_ar_scavenger1b1[selecao]

selecao = (valvula_dardo1_scavenger1b1.iloc[:,0] != 0) & (valvula_dardo1_scavenger1b1.iloc[:,0] != " ")
valvula_dardo1_scavenger1b1 = valvula_dardo1_scavenger1b1[selecao]

selecao = (valvula_dardo2_scavenger1b1.iloc[:,0] != 0) & (valvula_dardo2_scavenger1b1.iloc[:,0] != " ")
valvula_dardo2_scavenger1b1 = valvula_dardo2_scavenger1b1[selecao]

selecao = (valvula_ar_scavenger1b1.iloc[:,0] != 0) & (valvula_ar_scavenger1b1.iloc[:,0] != " ")
valvula_ar_scavenger1b1 = valvula_ar_scavenger1b1[selecao]

# Seleciona duas colunas, formando um agrupamento com os dados de um parâmetro
nivel_espuma_scavenger1b2 = scavenger1b2.iloc[:,[0,1]]
vazao_ar_scavenger1b2 = scavenger1b2.iloc[:,[2,3]]
valvula_dardo1_scavenger1b2 = scavenger1b2.iloc[:,[4,5]]
valvula_dardo2_scavenger1b2 = scavenger1b2.iloc[:,[6,7]]
valvula_ar_scavenger1b2 = scavenger1b2.iloc[:,[8,9]]

# Retira os dados nulos
selecao = (nivel_espuma_scavenger1b2.iloc[:,0] != 0) & (nivel_espuma_scavenger1b2.iloc[:,0] != " ")
nivel_espuma_scavenger1b2 = nivel_espuma_scavenger1b2[selecao]

selecao = (vazao_ar_scavenger1b2.iloc[:,0] != 0) & (vazao_ar_scavenger1b2.iloc[:,0] != " ")
vazao_ar_scavenger1b2 = vazao_ar_scavenger1b2[selecao]

selecao = (valvula_dardo1_scavenger1b2.iloc[:,0] != 0) & (valvula_dardo1_scavenger1b2.iloc[:,0] != " ")
valvula_dardo1_scavenger1b2 = valvula_dardo1_scavenger1b2[selecao]

selecao = (valvula_dardo2_scavenger1b2.iloc[:,0] != 0) & (valvula_dardo2_scavenger1b2.iloc[:,0] != " ")
valvula_dardo2_scavenger1b2 = valvula_dardo2_scavenger1b2[selecao]

selecao = (valvula_ar_scavenger1b2.iloc[:,0] != 0) & (valvula_ar_scavenger1b2.iloc[:,0] != " ")
valvula_ar_scavenger1b2 = valvula_ar_scavenger1b2[selecao]

# Seleciona duas colunas, formando um agrupamento com os dados de um parâmetro
nivel_espuma_scavenger1b3 = scavenger1b3.iloc[:,[0,1]]
vazao_ar_scavenger1b3 = scavenger1b3.iloc[:,[2,3]]
valvula_dardo1_scavenger1b3 = scavenger1b3.iloc[:,[4,5]]
valvula_dardo2_scavenger1b3 = scavenger1b3.iloc[:,[6,7]]
valvula_ar_scavenger1b3 = scavenger1b3.iloc[:,[8,9]]

# Retira os dados nulos
selecao = (nivel_espuma_scavenger1b3.iloc[:,0] != 0) & (nivel_espuma_scavenger1b3.iloc[:,0] != " ")
nivel_espuma_scavenger1b3 = nivel_espuma_scavenger1b3[selecao]

selecao = (vazao_ar_scavenger1b3.iloc[:,0] != 0) & (vazao_ar_scavenger1b3.iloc[:,0] != " ")
vazao_ar_scavenger1b3 = vazao_ar_scavenger1b3[selecao]

selecao = (valvula_dardo1_scavenger1b3.iloc[:,0] != 0) & (valvula_dardo1_scavenger1b3.iloc[:,0] != " ")
valvula_dardo1_scavenger1b3 = valvula_dardo1_scavenger1b3[selecao]

selecao = (valvula_dardo2_scavenger1b3.iloc[:,0] != 0) & (valvula_dardo2_scavenger1b3.iloc[:,0] != " ")
valvula_dardo2_scavenger1b3 = valvula_dardo2_scavenger1b3[selecao]

selecao = (valvula_ar_scavenger1b3.iloc[:,0] != 0) & (valvula_ar_scavenger1b3.iloc[:,0] != " ")
valvula_ar_scavenger1b3 = valvula_ar_scavenger1b3[selecao]

# Seleciona duas colunas, formando um agrupamento com os dados de um parâmetro
nivel_espuma_scavenger2b1 = scavenger2b1.iloc[:,[0,1]]
vazao_ar_scavenger2b1 = scavenger2b1.iloc[:,[2,3]]
valvula_dardo1_scavenger2b1 = scavenger2b1.iloc[:,[4,5]]
valvula_dardo2_scavenger2b1 = scavenger2b1.iloc[:,[6,7]]
valvula_ar_scavenger2b1 = scavenger2b1.iloc[:,[8,9]]

# Retira os dados nulos
selecao = (nivel_espuma_scavenger2b1.iloc[:,0] != 0) & (nivel_espuma_scavenger2b1.iloc[:,0] != " ")
nivel_espuma_scavenger2b1 = nivel_espuma_scavenger2b1[selecao]

selecao = (vazao_ar_scavenger2b1.iloc[:,0] != 0) & (vazao_ar_scavenger2b1.iloc[:,0] != " ")
vazao_ar_scavenger2b1 = vazao_ar_scavenger2b1[selecao]

selecao = (valvula_dardo1_scavenger2b1.iloc[:,0] != 0) & (valvula_dardo1_scavenger2b1.iloc[:,0] != " ")
valvula_dardo1_scavenger2b1 = valvula_dardo1_scavenger2b1[selecao]

selecao = (valvula_dardo2_scavenger2b1.iloc[:,0] != 0) & (valvula_dardo2_scavenger2b1.iloc[:,0] != " ")
valvula_dardo2_scavenger2b1 = valvula_dardo2_scavenger2b1[selecao]

selecao = (valvula_ar_scavenger2b1.iloc[:,0] != 0) & (valvula_ar_scavenger2b1.iloc[:,0] != " ")
valvula_ar_scavenger2b1 = valvula_ar_scavenger2b1[selecao]

# Seleciona duas colunas, formando um agrupamento com os dados de um parâmetro
nivel_espuma_scavenger2b2 = scavenger2b2.iloc[:,[0,1]]
vazao_ar_scavenger2b2 = scavenger2b2.iloc[:,[2,3]]
valvula_dardo1_scavenger2b2 = scavenger2b2.iloc[:,[4,5]]
valvula_dardo2_scavenger2b2 = scavenger2b2.iloc[:,[6,7]]
valvula_ar_scavenger2b2 = scavenger2b2.iloc[:,[8,9]]

# Retira os dados nulos
selecao = (nivel_espuma_scavenger2b2.iloc[:,0] != 0) & (nivel_espuma_scavenger2b2.iloc[:,0] != " ")
nivel_espuma_scavenger2b2 = nivel_espuma_scavenger2b2[selecao]

selecao = (vazao_ar_scavenger2b2.iloc[:,0] != 0) & (vazao_ar_scavenger2b2.iloc[:,0] != " ")
vazao_ar_scavenger2b2 = vazao_ar_scavenger2b2[selecao]

selecao = (valvula_dardo1_scavenger2b2.iloc[:,0] != 0) & (valvula_dardo1_scavenger2b2.iloc[:,0] != " ")
valvula_dardo1_scavenger2b2 = valvula_dardo1_scavenger2b2[selecao]

selecao = (valvula_dardo2_scavenger2b2.iloc[:,0] != 0) & (valvula_dardo2_scavenger2b2.iloc[:,0] != " ")
valvula_dardo2_scavenger2b2 = valvula_dardo2_scavenger2b2[selecao]

selecao = (valvula_ar_scavenger2b2.iloc[:,0] != 0) & (valvula_ar_scavenger2b2.iloc[:,0] != " ")
valvula_ar_scavenger2b2 = valvula_ar_scavenger2b2[selecao]

# Seleciona duas colunas, formando um agrupamento com os dados de um parâmetro
nivel_espuma_scavenger2b3 = scavenger2b3.iloc[:,[0,1]]
vazao_ar_scavenger2b3 = scavenger2b3.iloc[:,[2,3]]
valvula_dardo1_scavenger2b3 = scavenger2b3.iloc[:,[4,5]]
valvula_dardo2_scavenger2b3 = scavenger2b3.iloc[:,[6,7]]
valvula_ar_scavenger2b3 = scavenger2b3.iloc[:,[8,9]]

# Retira os dados nulos
selecao = (nivel_espuma_scavenger2b3.iloc[:,0] != 0) & (nivel_espuma_scavenger2b3.iloc[:,0] != " ")
nivel_espuma_scavenger2b3 = nivel_espuma_scavenger2b3[selecao]

selecao = (vazao_ar_scavenger2b3.iloc[:,0] != 0) & (vazao_ar_scavenger2b3.iloc[:,0] != " ")
vazao_ar_scavenger2b3 = vazao_ar_scavenger2b3[selecao]

selecao = (valvula_dardo1_scavenger2b3.iloc[:,0] != 0) & (valvula_dardo1_scavenger2b3.iloc[:,0] != " ")
valvula_dardo1_scavenger2b3 = valvula_dardo1_scavenger2b3[selecao]

selecao = (valvula_dardo2_scavenger2b3.iloc[:,0] != 0) & (valvula_dardo2_scavenger2b3.iloc[:,0] != " ")
valvula_dardo2_scavenger2b3 = valvula_dardo2_scavenger2b3[selecao]

selecao = (valvula_ar_scavenger2b3.iloc[:,0] != 0) & (valvula_ar_scavenger2b3.iloc[:,0] != " ")
valvula_ar_scavenger2b3 = valvula_ar_scavenger2b3[selecao]

# Remove os dados que estão fora do intervalo
selecao = (nivel_espuma_scavenger1b1.iloc[:,0] > base) & (nivel_espuma_scavenger1b1.iloc[:,0] <= fim)
nivel_espuma_scavenger1b1 = nivel_espuma_scavenger1b1[selecao]

selecao = (vazao_ar_scavenger1b1.iloc[:,0] > base) & (vazao_ar_scavenger1b1.iloc[:,0] <= fim)
vazao_ar_scavenger1b1 = vazao_ar_scavenger1b1[selecao]

selecao = (valvula_dardo1_scavenger1b1.iloc[:,0] > base) & (valvula_dardo1_scavenger1b1.iloc[:,0] <= fim)
valvula_dardo1_scavenger1b1 = valvula_dardo1_scavenger1b1[selecao]

selecao = (valvula_dardo2_scavenger1b1.iloc[:,0] > base) & (valvula_dardo2_scavenger1b1.iloc[:,0] <= fim)
valvula_dardo2_scavenger1b1 = valvula_dardo2_scavenger1b1[selecao]

selecao = (valvula_ar_scavenger1b1.iloc[:,0] > base) & (valvula_ar_scavenger1b1.iloc[:,0] <= fim)
valvula_ar_scavenger1b1 = valvula_ar_scavenger1b1[selecao]

# Remove os dados que estão fora do intervalo
selecao = (nivel_espuma_scavenger1b2.iloc[:,0] > base) & (nivel_espuma_scavenger1b2.iloc[:,0] <= fim)
nivel_espuma_scavenger1b2 = nivel_espuma_scavenger1b2[selecao]

selecao = (vazao_ar_scavenger1b2.iloc[:,0] > base) & (vazao_ar_scavenger1b2.iloc[:,0] <= fim)
vazao_ar_scavenger1b2 = vazao_ar_scavenger1b2[selecao]

selecao = (valvula_dardo1_scavenger1b2.iloc[:,0] > base) & (valvula_dardo1_scavenger1b2.iloc[:,0] <= fim)
valvula_dardo1_scavenger1b2 = valvula_dardo1_scavenger1b2[selecao]

selecao = (valvula_dardo2_scavenger1b2.iloc[:,0] > base) & (valvula_dardo2_scavenger1b2.iloc[:,0] <= fim)
valvula_dardo2_scavenger1b2 = valvula_dardo2_scavenger1b2[selecao]

selecao = (valvula_ar_scavenger1b2.iloc[:,0] > base) & (valvula_ar_scavenger1b2.iloc[:,0] <= fim)
valvula_ar_scavenger1b2 = valvula_ar_scavenger1b2[selecao]

# Remove os dados que estão fora do intervalo
selecao = (nivel_espuma_scavenger1b3.iloc[:,0] > base) & (nivel_espuma_scavenger1b3.iloc[:,0] <= fim)
nivel_espuma_scavenger1b3 = nivel_espuma_scavenger1b3[selecao]

selecao = (vazao_ar_scavenger1b3.iloc[:,0] > base) & (vazao_ar_scavenger1b3.iloc[:,0] <= fim)
vazao_ar_scavenger1b3 = vazao_ar_scavenger1b3[selecao]

selecao = (valvula_dardo1_scavenger1b3.iloc[:,0] > base) & (valvula_dardo1_scavenger1b3.iloc[:,0] <= fim)
valvula_dardo1_scavenger1b3 = valvula_dardo1_scavenger1b3[selecao]

selecao = (valvula_dardo2_scavenger1b3.iloc[:,0] > base) & (valvula_dardo2_scavenger1b3.iloc[:,0] <= fim)
valvula_dardo2_scavenger1b3 = valvula_dardo2_scavenger1b3[selecao]

selecao = (valvula_ar_scavenger1b3.iloc[:,0] > base) & (valvula_ar_scavenger1b3.iloc[:,0] <= fim)
valvula_ar_scavenger1b3 = valvula_ar_scavenger1b3[selecao]

# Remove os dados que estão fora do intervalo
selecao = (nivel_espuma_scavenger2b1.iloc[:,0] > base) & (nivel_espuma_scavenger2b1.iloc[:,0] <= fim)
nivel_espuma_scavenger2b1 = nivel_espuma_scavenger2b1[selecao]

selecao = (vazao_ar_scavenger2b1.iloc[:,0] > base) & (vazao_ar_scavenger2b1.iloc[:,0] <= fim)
vazao_ar_scavenger2b1 = vazao_ar_scavenger2b1[selecao]

selecao = (valvula_dardo1_scavenger2b1.iloc[:,0] > base) & (valvula_dardo1_scavenger2b1.iloc[:,0] <= fim)
valvula_dardo1_scavenger2b1 = valvula_dardo1_scavenger2b1[selecao]

selecao = (valvula_dardo2_scavenger2b1.iloc[:,0] > base) & (valvula_dardo2_scavenger2b1.iloc[:,0] <= fim)
valvula_dardo2_scavenger2b1 = valvula_dardo2_scavenger2b1[selecao]

selecao = (valvula_ar_scavenger2b1.iloc[:,0] > base) & (valvula_ar_scavenger2b1.iloc[:,0] <= fim)
valvula_ar_scavenger2b1 = valvula_ar_scavenger2b1[selecao]

# Remove os dados que estão fora do intervalo
selecao = (nivel_espuma_scavenger2b2.iloc[:,0] > base) & (nivel_espuma_scavenger2b2.iloc[:,0] <= fim)
nivel_espuma_scavenger2b2 = nivel_espuma_scavenger2b2[selecao]

selecao = (vazao_ar_scavenger2b2.iloc[:,0] > base) & (vazao_ar_scavenger2b2.iloc[:,0] <= fim)
vazao_ar_scavenger2b2 = vazao_ar_scavenger2b2[selecao]

selecao = (valvula_dardo1_scavenger2b2.iloc[:,0] > base) & (valvula_dardo1_scavenger2b2.iloc[:,0] <= fim)
valvula_dardo1_scavenger2b2 = valvula_dardo1_scavenger2b2[selecao]

selecao = (valvula_dardo2_scavenger2b2.iloc[:,0] > base) & (valvula_dardo2_scavenger2b2.iloc[:,0] <= fim)
valvula_dardo2_scavenger2b2 = valvula_dardo2_scavenger2b2[selecao]

selecao = (valvula_ar_scavenger2b2.iloc[:,0] > base) & (valvula_ar_scavenger2b2.iloc[:,0] <= fim)
valvula_ar_scavenger2b2 = valvula_ar_scavenger2b2[selecao]

# Remove os dados que estão fora do intervalo
selecao = (nivel_espuma_scavenger2b3.iloc[:,0] > base) & (nivel_espuma_scavenger2b3.iloc[:,0] <= fim)
nivel_espuma_scavenger2b3 = nivel_espuma_scavenger2b3[selecao]

selecao = (vazao_ar_scavenger2b3.iloc[:,0] > base) & (vazao_ar_scavenger2b3.iloc[:,0] <= fim)
vazao_ar_scavenger2b3 = vazao_ar_scavenger2b3[selecao]

selecao = (valvula_dardo1_scavenger2b3.iloc[:,0] > base) & (valvula_dardo1_scavenger2b3.iloc[:,0] <= fim)
valvula_dardo1_scavenger2b3 = valvula_dardo1_scavenger2b3[selecao]

selecao = (valvula_dardo2_scavenger2b3.iloc[:,0] > base) & (valvula_dardo2_scavenger2b3.iloc[:,0] <= fim)
valvula_dardo2_scavenger2b3 = valvula_dardo2_scavenger2b3[selecao]

selecao = (valvula_ar_scavenger2b3.iloc[:,0] > base) & (valvula_ar_scavenger2b3.iloc[:,0] <= fim)
valvula_ar_scavenger2b3 = valvula_ar_scavenger2b3[selecao]

# Zera o índice
#nivel_espuma_scavenger1b1.index = [j for j in range(nivel_espuma_scavenger1b1.iloc[:,0].shape[0])]
#vazao_ar_scavenger1b1.index = [j for j in range(vazao_ar_scavenger1b1.iloc[:,0].shape[0])]
#valvula_dardo1_scavenger1b1.index = [j for j in range(valvula_dardo1_scavenger1b1.iloc[:,0].shape[0])]
#valvula_dardo2_scavenger1b1.index = [j for j in range(valvula_dardo2_scavenger1b1.iloc[:,0].shape[0])]
#valvula_ar_scavenger1b1.index = [j for j in range(valvula_ar_scavenger1b1.iloc[:,0].shape[0])]
nivel_espuma_scavenger1b1.reset_index(drop=True, inplace=True)
vazao_ar_scavenger1b1.reset_index(drop=True, inplace=True)
valvula_dardo1_scavenger1b1.reset_index(drop=True, inplace=True)
valvula_dardo2_scavenger1b1.reset_index(drop=True, inplace=True)
valvula_ar_scavenger1b1.reset_index(drop=True, inplace=True)

# Zera o índice
#nivel_espuma_scavenger1b2.index = [j for j in range(nivel_espuma_scavenger1b2.iloc[:,0].shape[0])]
#vazao_ar_scavenger1b2.index = [j for j in range(vazao_ar_scavenger1b2.iloc[:,0].shape[0])]
#valvula_dardo1_scavenger1b2.index = [j for j in range(valvula_dardo1_scavenger1b2.iloc[:,0].shape[0])]
#valvula_dardo2_scavenger1b2.index = [j for j in range(valvula_dardo2_scavenger1b2.iloc[:,0].shape[0])]
#valvula_ar_scavenger1b2.index = [j for j in range(valvula_ar_scavenger1b2.iloc[:,0].shape[0])]
nivel_espuma_scavenger1b2.reset_index(drop=True, inplace=True)
vazao_ar_scavenger1b2.reset_index(drop=True, inplace=True)
valvula_dardo1_scavenger1b2.reset_index(drop=True, inplace=True)
valvula_dardo2_scavenger1b2.reset_index(drop=True, inplace=True)
valvula_ar_scavenger1b2.reset_index(drop=True, inplace=True)

# Zera o índice
#nivel_espuma_scavenger1b3.index = [j for j in range(nivel_espuma_scavenger1b3.iloc[:,0].shape[0])]
#vazao_ar_scavenger1b3.index = [j for j in range(vazao_ar_scavenger1b3.iloc[:,0].shape[0])]
#valvula_dardo1_scavenger1b3.index = [j for j in range(valvula_dardo1_scavenger1b3.iloc[:,0].shape[0])]
#valvula_dardo2_scavenger1b3.index = [j for j in range(valvula_dardo2_scavenger1b3.iloc[:,0].shape[0])]
#valvula_ar_scavenger1b3.index = [j for j in range(valvula_ar_scavenger1b3.iloc[:,0].shape[0])]
nivel_espuma_scavenger1b3.reset_index(drop=True, inplace=True)
vazao_ar_scavenger1b3.reset_index(drop=True, inplace=True)
valvula_dardo1_scavenger1b3.reset_index(drop=True, inplace=True)
valvula_dardo2_scavenger1b3.reset_index(drop=True, inplace=True)
valvula_ar_scavenger1b3.reset_index(drop=True, inplace=True)

# Zera o índice
#nivel_espuma_scavenger2b1.index = [j for j in range(nivel_espuma_scavenger2b1.iloc[:,0].shape[0])]
#vazao_ar_scavenger2b1.index = [j for j in range(vazao_ar_scavenger2b1.iloc[:,0].shape[0])]
#valvula_dardo1_scavenger2b1.index = [j for j in range(valvula_dardo1_scavenger2b1.iloc[:,0].shape[0])]
#valvula_dardo2_scavenger2b1.index = [j for j in range(valvula_dardo2_scavenger2b1.iloc[:,0].shape[0])]
#valvula_ar_scavenger2b1.index = [j for j in range(valvula_ar_scavenger2b1.iloc[:,0].shape[0])]
nivel_espuma_scavenger2b1.reset_index(drop=True, inplace=True)
vazao_ar_scavenger2b1.reset_index(drop=True, inplace=True)
valvula_dardo1_scavenger2b1.reset_index(drop=True, inplace=True)
valvula_dardo2_scavenger2b1.reset_index(drop=True, inplace=True)
valvula_ar_scavenger2b1.reset_index(drop=True, inplace=True)

# Zera o índice
#nivel_espuma_scavenger2b2.index = [j for j in range(nivel_espuma_scavenger2b2.iloc[:,0].shape[0])]
#vazao_ar_scavenger2b2.index = [j for j in range(vazao_ar_scavenger2b2.iloc[:,0].shape[0])]
#valvula_dardo1_scavenger2b2.index = [j for j in range(valvula_dardo1_scavenger2b2.iloc[:,0].shape[0])]
#valvula_dardo2_scavenger2b2.index = [j for j in range(valvula_dardo2_scavenger2b2.iloc[:,0].shape[0])]
#valvula_ar_scavenger2b2.index = [j for j in range(valvula_ar_scavenger2b2.iloc[:,0].shape[0])]
nivel_espuma_scavenger2b2.reset_index(drop=True, inplace=True)
vazao_ar_scavenger2b2.reset_index(drop=True, inplace=True)
valvula_dardo1_scavenger2b2.reset_index(drop=True, inplace=True)
valvula_dardo2_scavenger2b2.reset_index(drop=True, inplace=True)
valvula_ar_scavenger2b2.reset_index(drop=True, inplace=True)

# Zera o índice
#nivel_espuma_scavenger2b3.index = [j for j in range(nivel_espuma_scavenger2b3.iloc[:,0].shape[0])]
#vazao_ar_scavenger2b3.index = [j for j in range(vazao_ar_scavenger2b3.iloc[:,0].shape[0])]
#valvula_dardo1_scavenger2b3.index = [j for j in range(valvula_dardo1_scavenger2b3.iloc[:,0].shape[0])]
#valvula_dardo2_scavenger2b3.index = [j for j in range(valvula_dardo2_scavenger2b3.iloc[:,0].shape[0])]
#valvula_ar_scavenger2b3.index = [j for j in range(valvula_ar_scavenger2b3.iloc[:,0].shape[0])]
nivel_espuma_scavenger2b3.reset_index(drop=True, inplace=True)
vazao_ar_scavenger2b3.reset_index(drop=True, inplace=True)
valvula_dardo1_scavenger2b3.reset_index(drop=True, inplace=True)
valvula_dardo2_scavenger2b3.reset_index(drop=True, inplace=True)
valvula_ar_scavenger2b3.reset_index(drop=True, inplace=True)

# Cria coluna para separação em grupos
nivel_espuma_scavenger1b1['gr'] = -1
vazao_ar_scavenger1b1['gr'] = -1
valvula_dardo1_scavenger1b1['gr'] = -1
valvula_dardo2_scavenger1b1['gr'] = -1
valvula_ar_scavenger1b1['gr'] = -1

# Cria coluna para separação em grupos
nivel_espuma_scavenger1b2['gr'] = -1
vazao_ar_scavenger1b2['gr'] = -1
valvula_dardo1_scavenger1b2['gr'] = -1
valvula_dardo2_scavenger1b2['gr'] = -1
valvula_ar_scavenger1b2['gr'] = -1

# Cria coluna para separação em grupos
nivel_espuma_scavenger1b3['gr'] = -1
vazao_ar_scavenger1b3['gr'] = -1
valvula_dardo1_scavenger1b3['gr'] = -1
valvula_dardo2_scavenger1b3['gr'] = -1
valvula_ar_scavenger1b3['gr'] = -1

# Cria coluna para separação em grupos
nivel_espuma_scavenger2b1['gr'] = -1
vazao_ar_scavenger2b1['gr'] = -1
valvula_dardo1_scavenger2b1['gr'] = -1
valvula_dardo2_scavenger2b1['gr'] = -1
valvula_ar_scavenger2b1['gr'] = -1

# Cria coluna para separação em grupos
nivel_espuma_scavenger2b2['gr'] = -1
vazao_ar_scavenger2b2['gr'] = -1
valvula_dardo1_scavenger2b2['gr'] = -1
valvula_dardo2_scavenger2b2['gr'] = -1
valvula_ar_scavenger2b2['gr'] = -1

# Cria coluna para separação em grupos
nivel_espuma_scavenger2b3['gr'] = -1
vazao_ar_scavenger2b3['gr'] = -1
valvula_dardo1_scavenger2b3['gr'] = -1
valvula_dardo2_scavenger2b3['gr'] = -1
valvula_ar_scavenger2b3['gr'] = -1

i = 0
base = bkp_base
for amostra in target.iloc[:,0]:
    # Critério de seleção dos dados com base na data e hora
    selecao = (nivel_espuma_scavenger1b1.iloc[:,0] > base) & (nivel_espuma_scavenger1b1.iloc[:,0] <= amostra)
    
    nivel_espuma_scavenger1b1.loc[selecao, 'gr'] = i
    
    # Critério de seleção dos dados com base na data e hora
    selecao = (vazao_ar_scavenger1b1.iloc[:,0] > base) & (vazao_ar_scavenger1b1.iloc[:,0] <= amostra)
    
    vazao_ar_scavenger1b1.loc[selecao, 'gr'] = i
    
    # Critério de seleção dos dados com base na data e hora
    selecao = (valvula_dardo1_scavenger1b1.iloc[:,0] > base) & (valvula_dardo1_scavenger1b1.iloc[:,0] <= amostra)
    
    valvula_dardo1_scavenger1b1.loc[selecao, 'gr'] = i
    
    # Critério de seleção dos dados com base na data e hora
    selecao = (valvula_dardo2_scavenger1b1.iloc[:,0] > base) & (valvula_dardo2_scavenger1b1.iloc[:,0] <= amostra)
    
    valvula_dardo2_scavenger1b1.loc[selecao, 'gr'] = i
    
    # Critério de seleção dos dados com base na data e hora
    selecao = (valvula_ar_scavenger1b1.iloc[:,0] > base) & (valvula_ar_scavenger1b1.iloc[:,0] <= amostra)
    
    valvula_ar_scavenger1b1.loc[selecao, 'gr'] = i
    
    base = amostra
    if i % 5 == 0:
        print(i)
    i += 1

i = 0
base = bkp_base
for amostra in target.iloc[:,0]:
    # Critério de seleção dos dados com base na data e hora
    selecao = (nivel_espuma_scavenger1b2.iloc[:,0] > base) & (nivel_espuma_scavenger1b2.iloc[:,0] <= amostra)
    
    nivel_espuma_scavenger1b2.loc[selecao, 'gr'] = i
    
    # Critério de seleção dos dados com base na data e hora
    selecao = (vazao_ar_scavenger1b2.iloc[:,0] > base) & (vazao_ar_scavenger1b2.iloc[:,0] <= amostra)
    
    vazao_ar_scavenger1b2.loc[selecao, 'gr'] = i
    
    # Critério de seleção dos dados com base na data e hora
    selecao = (valvula_dardo1_scavenger1b2.iloc[:,0] > base) & (valvula_dardo1_scavenger1b2.iloc[:,0] <= amostra)
    
    valvula_dardo1_scavenger1b2.loc[selecao, 'gr'] = i
    
    # Critério de seleção dos dados com base na data e hora
    selecao = (valvula_dardo2_scavenger1b2.iloc[:,0] > base) & (valvula_dardo2_scavenger1b2.iloc[:,0] <= amostra)
    
    valvula_dardo2_scavenger1b2.loc[selecao, 'gr'] = i
    
    # Critério de seleção dos dados com base na data e hora
    selecao = (valvula_ar_scavenger1b2.iloc[:,0] > base) & (valvula_ar_scavenger1b2.iloc[:,0] <= amostra)
    
    valvula_ar_scavenger1b2.loc[selecao, 'gr'] = i
    
    base = amostra
    if i % 5 == 0:
        print(i)
    i += 1

i = 0
base = bkp_base
for amostra in target.iloc[:,0]:
    # Critério de seleção dos dados com base na data e hora
    selecao = (nivel_espuma_scavenger1b3.iloc[:,0] > base) & (nivel_espuma_scavenger1b3.iloc[:,0] <= amostra)
    
    nivel_espuma_scavenger1b3.loc[selecao, 'gr'] = i
    
    # Critério de seleção dos dados com base na data e hora
    selecao = (vazao_ar_scavenger1b3.iloc[:,0] > base) & (vazao_ar_scavenger1b3.iloc[:,0] <= amostra)
    
    vazao_ar_scavenger1b3.loc[selecao, 'gr'] = i
    
    # Critério de seleção dos dados com base na data e hora
    selecao = (valvula_dardo1_scavenger1b3.iloc[:,0] > base) & (valvula_dardo1_scavenger1b3.iloc[:,0] <= amostra)
    
    valvula_dardo1_scavenger1b3.loc[selecao, 'gr'] = i
    
    # Critério de seleção dos dados com base na data e hora
    selecao = (valvula_dardo2_scavenger1b3.iloc[:,0] > base) & (valvula_dardo2_scavenger1b3.iloc[:,0] <= amostra)
    
    valvula_dardo2_scavenger1b3.loc[selecao, 'gr'] = i
    
    # Critério de seleção dos dados com base na data e hora
    selecao = (valvula_ar_scavenger1b3.iloc[:,0] > base) & (valvula_ar_scavenger1b3.iloc[:,0] <= amostra)
    
    valvula_ar_scavenger1b3.loc[selecao, 'gr'] = i
    
    base = amostra
    if i % 5 == 0:
        print(i)
    i += 1

i = 0
base = bkp_base
for amostra in target.iloc[:,0]:
    # Critério de seleção dos dados com base na data e hora
    selecao = (nivel_espuma_scavenger2b1.iloc[:,0] > base) & (nivel_espuma_scavenger2b1.iloc[:,0] <= amostra)
    
    nivel_espuma_scavenger2b1.loc[selecao, 'gr'] = i
    
    # Critério de seleção dos dados com base na data e hora
    selecao = (vazao_ar_scavenger2b1.iloc[:,0] > base) & (vazao_ar_scavenger2b1.iloc[:,0] <= amostra)
    
    vazao_ar_scavenger2b1.loc[selecao, 'gr'] = i
    
    # Critério de seleção dos dados com base na data e hora
    selecao = (valvula_dardo1_scavenger2b1.iloc[:,0] > base) & (valvula_dardo1_scavenger2b1.iloc[:,0] <= amostra)
    
    valvula_dardo1_scavenger2b1.loc[selecao, 'gr'] = i
    
    # Critério de seleção dos dados com base na data e hora
    selecao = (valvula_dardo2_scavenger2b1.iloc[:,0] > base) & (valvula_dardo2_scavenger2b1.iloc[:,0] <= amostra)
    
    valvula_dardo2_scavenger2b1.loc[selecao, 'gr'] = i
    
    # Critério de seleção dos dados com base na data e hora
    selecao = (valvula_ar_scavenger2b1.iloc[:,0] > base) & (valvula_ar_scavenger2b1.iloc[:,0] <= amostra)
    
    valvula_ar_scavenger2b1.loc[selecao, 'gr'] = i
    
    base = amostra
    if i % 5 == 0:
        print(i)
    i += 1

i = 0
base = bkp_base
for amostra in target.iloc[:,0]:
    # Critério de seleção dos dados com base na data e hora
    selecao = (nivel_espuma_scavenger2b2.iloc[:,0] > base) & (nivel_espuma_scavenger2b2.iloc[:,0] <= amostra)
    
    nivel_espuma_scavenger2b2.loc[selecao, 'gr'] = i
    
    # Critério de seleção dos dados com base na data e hora
    selecao = (vazao_ar_scavenger2b2.iloc[:,0] > base) & (vazao_ar_scavenger2b2.iloc[:,0] <= amostra)
    
    vazao_ar_scavenger2b2.loc[selecao, 'gr'] = i
    
    # Critério de seleção dos dados com base na data e hora
    selecao = (valvula_dardo1_scavenger2b2.iloc[:,0] > base) & (valvula_dardo1_scavenger2b2.iloc[:,0] <= amostra)
    
    valvula_dardo1_scavenger2b2.loc[selecao, 'gr'] = i
    
    # Critério de seleção dos dados com base na data e hora
    selecao = (valvula_dardo2_scavenger2b2.iloc[:,0] > base) & (valvula_dardo2_scavenger2b2.iloc[:,0] <= amostra)
    
    valvula_dardo2_scavenger2b2.loc[selecao, 'gr'] = i
    
    # Critério de seleção dos dados com base na data e hora
    selecao = (valvula_ar_scavenger2b2.iloc[:,0] > base) & (valvula_ar_scavenger2b2.iloc[:,0] <= amostra)
    
    valvula_ar_scavenger2b2.loc[selecao, 'gr'] = i
    
    base = amostra
    if i % 5 == 0:
        print(i)
    i += 1

i = 0
base = bkp_base
for amostra in target.iloc[:,0]:
    # Critério de seleção dos dados com base na data e hora
    selecao = (nivel_espuma_scavenger2b3.iloc[:,0] > base) & (nivel_espuma_scavenger2b3.iloc[:,0] <= amostra)
    
    nivel_espuma_scavenger2b3.loc[selecao, 'gr'] = i
    
    # Critério de seleção dos dados com base na data e hora
    selecao = (vazao_ar_scavenger2b3.iloc[:,0] > base) & (vazao_ar_scavenger2b3.iloc[:,0] <= amostra)
    
    vazao_ar_scavenger2b3.loc[selecao, 'gr'] = i
    
    # Critério de seleção dos dados com base na data e hora
    selecao = (valvula_dardo1_scavenger2b3.iloc[:,0] > base) & (valvula_dardo1_scavenger2b3.iloc[:,0] <= amostra)
    
    valvula_dardo1_scavenger2b3.loc[selecao, 'gr'] = i
    
    # Critério de seleção dos dados com base na data e hora
    selecao = (valvula_dardo2_scavenger2b3.iloc[:,0] > base) & (valvula_dardo2_scavenger2b3.iloc[:,0] <= amostra)
    
    valvula_dardo2_scavenger2b3.loc[selecao, 'gr'] = i
    
    # Critério de seleção dos dados com base na data e hora
    selecao = (valvula_ar_scavenger2b3.iloc[:,0] > base) & (valvula_ar_scavenger2b3.iloc[:,0] <= amostra)
    
    valvula_ar_scavenger2b3.loc[selecao, 'gr'] = i
    
    base = amostra
    if i % 5 == 0:
        print(i)
    i += 1

nivel_espuma_scavenger1b1.to_excel(pasta + 'S1B1_Nivel_Espuma.xlsx', 'Planilha1', index=False)

vazao_ar_scavenger1b1.to_excel(pasta + 'S1B1_Vazao_Ar.xlsx', 'Planilha1', index=False)

valvula_dardo1_scavenger1b1.to_excel(pasta + 'S1B1_Valv_Dardo1.xlsx', 'Planilha1', index=False)

valvula_dardo2_scavenger1b1.to_excel(pasta + 'S1B1_Valv_Dardo2.xlsx', 'Planilha1', index=False)

valvula_ar_scavenger1b1.to_excel(pasta + 'S1B1_Valv_Ar.xlsx', 'Planilha1', index=False)

nivel_espuma_scavenger1b2.to_excel(pasta + 'S1B2_Nivel_Espuma.xlsx', 'Planilha1', index=False)

vazao_ar_scavenger1b2.to_excel(pasta + 'S1B2_Vazao_Ar.xlsx', 'Planilha1', index=False)

valvula_dardo1_scavenger1b2.to_excel(pasta + 'S1B2_Valv_Dardo1.xlsx', 'Planilha1', index=False)

valvula_dardo2_scavenger1b2.to_excel(pasta + 'S1B2_Valv_Dardo2.xlsx', 'Planilha1', index=False)

valvula_ar_scavenger1b2.to_excel(pasta + 'S1B2_Valv_Ar.xlsx', 'Planilha1', index=False)

nivel_espuma_scavenger1b3.to_excel(pasta + 'S1B3_Nivel_Espuma.xlsx', 'Planilha1', index=False)

vazao_ar_scavenger1b3.to_excel(pasta + 'S1B3_Vazao_Ar.xlsx', 'Planilha1', index=False)

valvula_dardo1_scavenger1b3.to_excel(pasta + 'S1B3_Valv_Dardo1.xlsx', 'Planilha1', index=False)

valvula_dardo2_scavenger1b3.to_excel(pasta + 'S1B3_Valv_Dardo2.xlsx', 'Planilha1', index=False)

valvula_ar_scavenger1b3.to_excel(pasta + 'S1B3_Valv_Ar.xlsx', 'Planilha1', index=False)

nivel_espuma_scavenger2b1.to_excel(pasta + 'S2B1_Nivel_Espuma.xlsx', 'Planilha1', index=False)

vazao_ar_scavenger2b1.to_excel(pasta + 'S2B1_Vazao_Ar.xlsx', 'Planilha1', index=False)

valvula_dardo1_scavenger2b1.to_excel(pasta + 'S2B1_Valv_Dardo1.xlsx', 'Planilha1', index=False)

valvula_dardo2_scavenger2b1.to_excel(pasta + 'S2B1_Valv_Dardo2.xlsx', 'Planilha1', index=False)

valvula_ar_scavenger2b1.to_excel(pasta + 'S2B1_Valv_Ar.xlsx', 'Planilha1', index=False)

nivel_espuma_scavenger2b2.to_excel(pasta + 'S2B2_Nivel_Espuma.xlsx', 'Planilha1', index=False)

vazao_ar_scavenger2b2.to_excel(pasta + 'S2B2_Vazao_Ar.xlsx', 'Planilha1', index=False)

valvula_dardo1_scavenger2b2.to_excel(pasta + 'S2B2_Valv_Dardo1.xlsx', 'Planilha1', index=False)

valvula_dardo2_scavenger2b2.to_excel(pasta + 'S2B2_Valv_Dardo2.xlsx', 'Planilha1', index=False)

valvula_ar_scavenger2b2.to_excel(pasta + 'S2B2_Valv_Ar.xlsx', 'Planilha1', index=False)

nivel_espuma_scavenger2b3.to_excel(pasta + 'S2B3_Nivel_Espuma.xlsx', 'Planilha1', index=False)

vazao_ar_scavenger2b3.to_excel(pasta + 'S2B3_Vazao_Ar.xlsx', 'Planilha1', index=False)

valvula_dardo1_scavenger2b3.to_excel(pasta + 'S2B3_Valv_Dardo1.xlsx', 'Planilha1', index=False)

valvula_dardo2_scavenger2b3.to_excel(pasta + 'S2B3_Valv_Dardo2.xlsx', 'Planilha1', index=False)

valvula_ar_scavenger2b3.to_excel(pasta + 'S2B3_Valv_Ar.xlsx', 'Planilha1', index=False)
