==> 04-Agrupamentos.ipynb <==
# Importa as bibliotecas
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
import numpy as np

# Faz a leitura dos dados
target = pd.read_excel('Abril01/T_Conceicao_0401-0701.xlsx')


nivel_cx_alimentacao = pd.read_excel('Abril01/CA_Nivel_0401-0701.xlsx')

densidade_cx_alimentacao = pd.read_excel('Abril01/CA_Densidade_0401-0701.xlsx')

vazao_vol_cx_alimentacao = pd.read_excel('Abril01/CA_Vazao_Vol_0401-0701.xlsx')

vazao_mas_cx_alimentacao = pd.read_excel('Abril01/CA_Vazao_Mas_0401-0701.xlsx')

vazao_amina_cx_alimentacao = pd.read_excel('Abril01/CA_Vazao_Amina_0401-0701.xlsx')

vazao_amido_cx_alimentacao = pd.read_excel('Abril01/CA_Vazao_Amido_0401-0701.xlsx')

lista = [nivel_cx_alimentacao, densidade_cx_alimentacao, vazao_vol_cx_alimentacao,
         vazao_mas_cx_alimentacao, vazao_amina_cx_alimentacao, vazao_amido_cx_alimentacao]

nivel_cx_recirculacao = pd.read_excel('Abril01/CR_Nivel_0401-0701.xlsx')

densidade_cx_recirculacao = pd.read_excel('Abril01/CR_Densidade_0401-0701.xlsx')

vazao_vol_cx_recirculacao = pd.read_excel('Abril01/CR_Vazao_Vol_0401-0701.xlsx')

vazao_mas_cx_recirculacao = pd.read_excel('Abril01/CR_Vazao_Mas_0401-0701.xlsx')

lista.append(nivel_cx_recirculacao)
lista.append(densidade_cx_recirculacao)
lista.append(vazao_vol_cx_recirculacao)
lista.append(vazao_mas_cx_recirculacao)

nivel_espuma_rougher1 = pd.read_excel('Abril01/R1_Nivel_Espuma_0401-0701.xlsx')

vazao_ar_rougher1 = pd.read_excel('Abril01/R1_Vazao_Ar_0401-0701.xlsx')

valvula_dardo1_rougher1 = pd.read_excel('Abril01/R1_Valv_Dardo1_0401-0701.xlsx')

valvula_dardo2_rougher1 = pd.read_excel('Abril01/R1_Valv_Dardo2_0401-0701.xlsx')

valvula_ar_rougher1 = pd.read_excel('Abril01/R1_Valv_Ar_0401-0701.xlsx')

lista.append(nivel_espuma_rougher1)
lista.append(vazao_ar_rougher1)
lista.append(valvula_dardo1_rougher1)
lista.append(valvula_dardo2_rougher1)
lista.append(valvula_ar_rougher1)

nivel_espuma_rougher2 = pd.read_excel('Abril01/R2_Nivel_Espuma_0401-0701.xlsx')

vazao_ar_rougher2 = pd.read_excel('Abril01/R2_Vazao_Ar_0401-0701.xlsx')

valvula_dardo1_rougher2 = pd.read_excel('Abril01/R2_Valv_Dardo1_0401-0701.xlsx')

valvula_dardo2_rougher2 = pd.read_excel('Abril01/R2_Valv_Dardo2_0401-0701.xlsx')

valvula_ar_rougher2 = pd.read_excel('Abril01/R2_Valv_Ar_0401-0701.xlsx')

lista.append(nivel_espuma_rougher2)
lista.append(vazao_ar_rougher2)
lista.append(valvula_dardo1_rougher2)
lista.append(valvula_dardo2_rougher2)
lista.append(valvula_ar_rougher2)

nivel_espuma_cleaner1 = pd.read_excel('Abril01/C1_Nivel_Espuma_0401-0701.xlsx')

vazao_ar_cleaner1 = pd.read_excel('Abril01/C1_Vazao_Ar_0401-0701.xlsx')

valvula_dardo1_cleaner1 = pd.read_excel('Abril01/C1_Valv_Dardo1_0401-0701.xlsx')

valvula_dardo2_cleaner1 = pd.read_excel('Abril01/C1_Valv_Dardo2_0401-0701.xlsx')

valvula_ar_cleaner1 = pd.read_excel('Abril01/C1_Valv_Ar_0401-0701.xlsx')

lista.append(nivel_espuma_cleaner1)
lista.append(vazao_ar_cleaner1)
lista.append(valvula_dardo1_cleaner1)
lista.append(valvula_dardo2_cleaner1)
lista.append(valvula_ar_cleaner1)

nivel_espuma_cleaner2 = pd.read_excel('Abril01/C2_Nivel_Espuma_0401-0701.xlsx')

vazao_ar_cleaner2 = pd.read_excel('Abril01/C2_Vazao_Ar_0401-0701.xlsx')

valvula_dardo1_cleaner2 = pd.read_excel('Abril01/C2_Valv_Dardo1_0401-0701.xlsx')

valvula_dardo2_cleaner2 = pd.read_excel('Abril01/C2_Valv_Dardo2_0401-0701.xlsx')

valvula_ar_cleaner2 = pd.read_excel('Abril01/C2_Valv_Ar_0401-0701.xlsx')

lista.append(nivel_espuma_cleaner2)
lista.append(vazao_ar_cleaner2)
lista.append(valvula_dardo1_cleaner2)
lista.append(valvula_dardo2_cleaner2)
lista.append(valvula_ar_cleaner2)

nivel_espuma_recleaner = pd.read_excel('Abril01/RC_Nivel_Espuma_0401-0701.xlsx')

vazao_ar_recleaner = pd.read_excel('Abril01/RC_Vazao_Ar_0401-0701.xlsx')

valvula_dardo1_recleaner = pd.read_excel('Abril01/RC_Valv_Dardo1_0401-0701.xlsx')

valvula_dardo2_recleaner = pd.read_excel('Abril01/RC_Valv_Dardo2_0401-0701.xlsx')

valvula_ar_recleaner = pd.read_excel('Abril01/RC_Valv_Ar_0401-0701.xlsx')

lista.append(nivel_espuma_recleaner)
lista.append(vazao_ar_recleaner)
lista.append(valvula_dardo1_recleaner)
lista.append(valvula_dardo2_recleaner)
lista.append(valvula_ar_recleaner)

nivel_espuma_scavenger1b1 = pd.read_excel('Abril01/S1B1_Nivel_Espuma_0401-0701.xlsx')

vazao_ar_scavenger1b1 = pd.read_excel('Abril01/S1B1_Vazao_Ar_0401-0701.xlsx')

valvula_dardo1_scavenger1b1 = pd.read_excel('Abril01/S1B1_Valv_Dardo1_0401-0701.xlsx')

valvula_dardo2_scavenger1b1 = pd.read_excel('Abril01/S1B1_Valv_Dardo2_0401-0701.xlsx')

valvula_ar_scavenger1b1 = pd.read_excel('Abril01/S1B1_Valv_Ar_0401-0701.xlsx')

lista.append(nivel_espuma_scavenger1b1)
lista.append(vazao_ar_scavenger1b1)
lista.append(valvula_dardo1_scavenger1b1)
lista.append(valvula_dardo2_scavenger1b1)
lista.append(valvula_ar_scavenger1b1)

nivel_espuma_scavenger1b2 = pd.read_excel('Abril01/S1B2_Nivel_Espuma_0401-0701.xlsx')

vazao_ar_scavenger1b2 = pd.read_excel('Abril01/S1B2_Vazao_Ar_0401-0701.xlsx')

valvula_dardo1_scavenger1b2 = pd.read_excel('Abril01/S1B2_Valv_Dardo1_0401-0701.xlsx')

valvula_dardo2_scavenger1b2 = pd.read_excel('Abril01/S1B2_Valv_Dardo2_0401-0701.xlsx')

valvula_ar_scavenger1b2 = pd.read_excel('Abril01/S1B2_Valv_Ar_0401-0701.xlsx')

lista.append(nivel_espuma_scavenger1b2)
lista.append(vazao_ar_scavenger1b2)
lista.append(valvula_dardo1_scavenger1b2)
lista.append(valvula_dardo2_scavenger1b2)
lista.append(valvula_ar_scavenger1b2)

nivel_espuma_scavenger1b3 = pd.read_excel('Abril01/S1B3_Nivel_Espuma_0401-0701.xlsx')

vazao_ar_scavenger1b3 = pd.read_excel('Abril01/S1B3_Vazao_Ar_0401-0701.xlsx')

valvula_dardo1_scavenger1b3 = pd.read_excel('Abril01/S1B3_Valv_Dardo1_0401-0701.xlsx')

valvula_dardo2_scavenger1b3 = pd.read_excel('Abril01/S1B3_Valv_Dardo2_0401-0701.xlsx')

valvula_ar_scavenger1b3 = pd.read_excel('Abril01/S1B3_Valv_Ar_0401-0701.xlsx')

lista.append(nivel_espuma_scavenger1b3)
lista.append(vazao_ar_scavenger1b3)
lista.append(valvula_dardo1_scavenger1b3)
lista.append(valvula_dardo2_scavenger1b3)
lista.append(valvula_ar_scavenger1b3)

nivel_espuma_scavenger2b1 = pd.read_excel('Abril01/S2B1_Nivel_Espuma_0401-0701.xlsx')

vazao_ar_scavenger2b1 = pd.read_excel('Abril01/S2B1_Vazao_Ar_0401-0701.xlsx')

valvula_dardo1_scavenger2b1 = pd.read_excel('Abril01/S2B1_Valv_Dardo1_0401-0701.xlsx')

valvula_dardo2_scavenger2b1 = pd.read_excel('Abril01/S2B1_Valv_Dardo2_0401-0701.xlsx')

valvula_ar_scavenger2b1 = pd.read_excel('Abril01/S2B1_Valv_Ar_0401-0701.xlsx')

lista.append(nivel_espuma_scavenger2b1)
lista.append(vazao_ar_scavenger2b1)
lista.append(valvula_dardo1_scavenger2b1)
lista.append(valvula_dardo2_scavenger2b1)
lista.append(valvula_ar_scavenger2b1)

nivel_espuma_scavenger2b2 = pd.read_excel('Abril01/S2B2_Nivel_Espuma_0401-0701.xlsx')

vazao_ar_scavenger2b2 = pd.read_excel('Abril01/S2B2_Vazao_Ar_0401-0701.xlsx')

valvula_dardo1_scavenger2b2 = pd.read_excel('Abril01/S2B2_Valv_Dardo1_0401-0701.xlsx')

valvula_dardo2_scavenger2b2 = pd.read_excel('Abril01/S2B2_Valv_Dardo2_0401-0701.xlsx')

valvula_ar_scavenger2b2 = pd.read_excel('Abril01/S2B2_Valv_Ar_0401-0701.xlsx')

lista.append(nivel_espuma_scavenger2b2)
lista.append(vazao_ar_scavenger2b2)
lista.append(valvula_dardo1_scavenger2b2)
lista.append(valvula_dardo2_scavenger2b2)
lista.append(valvula_ar_scavenger2b2)

nivel_espuma_scavenger2b3 = pd.read_excel('Abril01/S2B3_Nivel_Espuma_0401-0701.xlsx')

vazao_ar_scavenger2b3 = pd.read_excel('Abril01/S2B3_Vazao_Ar_0401-0701.xlsx')

valvula_dardo1_scavenger2b3 = pd.read_excel('Abril01/S2B3_Valv_Dardo1_0401-0701.xlsx')

valvula_dardo2_scavenger2b3 = pd.read_excel('Abril01/S2B3_Valv_Dardo2_0401-0701.xlsx')

valvula_ar_scavenger2b3 = pd.read_excel('Abril01/S2B3_Valv_Ar_0401-0701.xlsx')

lista.append(nivel_espuma_scavenger2b3)
lista.append(vazao_ar_scavenger2b3)
lista.append(valvula_dardo1_scavenger2b3)
lista.append(valvula_dardo2_scavenger2b3)
lista.append(valvula_ar_scavenger2b3)

agrupados = pd.DataFrame(index=target.index)
for item in lista:
    modelo = pd.DataFrame(index=target.index)
    grupo = item.groupby(['gr']).mean()
    for linha in grupo.index:
        modelo.loc[linha, grupo.columns[0]] = grupo.loc[linha, grupo.columns[0]]
    agrupados = pd.concat([agrupados, modelo], axis = 1)
agrupados.drop(index=-1, inplace=True)

modelo = pd.DataFrame(index=target.index)

grupo = nivel_cx_alimentacao.groupby(['gr']).mean()
for linha in grupo.index:
    modelo.loc[linha, grupo.columns[0]] = grupo.loc[linha, grupo.columns[0]]
nivel_cx_alimentacao_g = modelo

modelo = pd.DataFrame(index=target.index)

grupo = densidade_cx_alimentacao.groupby(['gr']).mean()
for linha in grupo.index:
    modelo.loc[linha, grupo.columns[0]] = grupo.loc[linha, grupo.columns[0]]
densidade_cx_alimentacao_g = modelo

modelo = pd.DataFrame(index=target.index)

grupo = vazao_vol_cx_alimentacao.groupby(['gr']).mean()
for linha in grupo.index:
    modelo.loc[linha, grupo.columns[0]] = grupo.loc[linha, grupo.columns[0]]
vazao_vol_cx_alimentacao_g = modelo

modelo = pd.DataFrame(index=target.index)

grupo = vazao_mas_cx_alimentacao.groupby(['gr']).mean()
for linha in grupo.index:
    modelo.loc[linha, grupo.columns[0]] = grupo.loc[linha, grupo.columns[0]]
vazao_mas_cx_alimentacao_g = modelo

modelo = pd.DataFrame(index=target.index)

grupo = vazao_amina_cx_alimentacao.groupby(['gr']).mean()
for linha in grupo.index:
    modelo.loc[linha, grupo.columns[0]] = grupo.loc[linha, grupo.columns[0]]
vazao_amina_cx_alimentacao_g = modelo

modelo = pd.DataFrame(index=target.index)

grupo = vazao_amido_cx_alimentacao.groupby(['gr']).mean()
for linha in grupo.index:
    modelo.loc[linha, grupo.columns[0]] = grupo.loc[linha, grupo.columns[0]]
vazao_amido_cx_alimentacao_g = modelo

modelo = pd.DataFrame(index=target.index)

grupo = nivel_cx_recirculacao.groupby(['gr']).mean()
for linha in grupo.index:
    modelo.loc[linha, grupo.columns[0]] = grupo.loc[linha, grupo.columns[0]]
nivel_cx_recirculacao_g = modelo

modelo = pd.DataFrame(index=target.index)

grupo = densidade_cx_recirculacao.groupby(['gr']).mean()
for linha in grupo.index:
    modelo.loc[linha, grupo.columns[0]] = grupo.loc[linha, grupo.columns[0]]
densidade_cx_recirculacao_g = modelo

modelo = pd.DataFrame(index=target.index)

grupo = vazao_vol_cx_recirculacao.groupby(['gr']).mean()
for linha in grupo.index:
    modelo.loc[linha, grupo.columns[0]] = grupo.loc[linha, grupo.columns[0]]
vazao_vol_cx_recirculacao_g = modelo

modelo = pd.DataFrame(index=target.index)

grupo = vazao_mas_cx_recirculacao.groupby(['gr']).mean()
for linha in grupo.index:
    modelo.loc[linha, grupo.columns[0]] = grupo.loc[linha, grupo.columns[0]]
vazao_mas_cx_recirculacao_g = modelo

modelo = pd.DataFrame(index=target.index)

grupo = nivel_espuma_rougher1.groupby(['gr']).mean()
for linha in grupo.index:
    modelo.loc[linha, grupo.columns[0]] = grupo.loc[linha, grupo.columns[0]]
nivel_espuma_rougher1_g = modelo

modelo = pd.DataFrame(index=target.index)

grupo = vazao_ar_rougher1.groupby(['gr']).mean()
for linha in grupo.index:
    modelo.loc[linha, grupo.columns[0]] = grupo.loc[linha, grupo.columns[0]]
vazao_ar_rougher1_g = modelo

modelo = pd.DataFrame(index=target.index)

grupo = valvula_dardo1_rougher1.groupby(['gr']).mean()
for linha in grupo.index:
    modelo.loc[linha, grupo.columns[0]] = grupo.loc[linha, grupo.columns[0]]
valvula_dardo1_rougher1_g = modelo

modelo = pd.DataFrame(index=target.index)

grupo = valvula_dardo2_rougher1.groupby(['gr']).mean()
for linha in grupo.index:
    modelo.loc[linha, grupo.columns[0]] = grupo.loc[linha, grupo.columns[0]]
valvula_dardo2_rougher1_g = modelo

modelo = pd.DataFrame(index=target.index)

grupo = valvula_ar_rougher1.groupby(['gr']).mean()
for linha in grupo.index:
    modelo.loc[linha, grupo.columns[0]] = grupo.loc[linha, grupo.columns[0]]
valvula_ar_rougher1_g = modelo

modelo = pd.DataFrame(index=target.index)

grupo = nivel_espuma_rougher2.groupby(['gr']).mean()
for linha in grupo.index:
    modelo.loc[linha, grupo.columns[0]] = grupo.loc[linha, grupo.columns[0]]
nivel_espuma_rougher2_g = modelo

modelo = pd.DataFrame(index=target.index)

grupo = vazao_ar_rougher2.groupby(['gr']).mean()
for linha in grupo.index:
    modelo.loc[linha, grupo.columns[0]] = grupo.loc[linha, grupo.columns[0]]
vazao_ar_rougher2_g = modelo

modelo = pd.DataFrame(index=target.index)

grupo = valvula_dardo1_rougher2.groupby(['gr']).mean()
for linha in grupo.index:
    modelo.loc[linha, grupo.columns[0]] = grupo.loc[linha, grupo.columns[0]]
valvula_dardo1_rougher2_g = modelo

modelo = pd.DataFrame(index=target.index)

grupo = valvula_dardo2_rougher2.groupby(['gr']).mean()
for linha in grupo.index:
    modelo.loc[linha, grupo.columns[0]] = grupo.loc[linha, grupo.columns[0]]
valvula_dardo2_rougher2_g = modelo

modelo = pd.DataFrame(index=target.index)

grupo = valvula_ar_rougher2.groupby(['gr']).mean()
for linha in grupo.index:
    modelo.loc[linha, grupo.columns[0]] = grupo.loc[linha, grupo.columns[0]]
valvula_ar_rougher2_g = modelo

modelo = pd.DataFrame(index=target.index)

grupo = nivel_espuma_cleaner1.groupby(['gr']).mean()
for linha in grupo.index:
    modelo.loc[linha, grupo.columns[0]] = grupo.loc[linha, grupo.columns[0]]
nivel_espuma_cleaner1_g = modelo

modelo = pd.DataFrame(index=target.index)

grupo = vazao_ar_cleaner1.groupby(['gr']).mean()
for linha in grupo.index:
    modelo.loc[linha, grupo.columns[0]] = grupo.loc[linha, grupo.columns[0]]
vazao_ar_cleaner1_g = modelo

modelo = pd.DataFrame(index=target.index)

grupo = valvula_dardo1_cleaner1.groupby(['gr']).mean()
for linha in grupo.index:
    modelo.loc[linha, grupo.columns[0]] = grupo.loc[linha, grupo.columns[0]]
valvula_dardo1_cleaner1_g = modelo

modelo = pd.DataFrame(index=target.index)

grupo = valvula_dardo2_cleaner1.groupby(['gr']).mean()
for linha in grupo.index:
    modelo.loc[linha, grupo.columns[0]] = grupo.loc[linha, grupo.columns[0]]
valvula_dardo2_cleaner1_g = modelo

modelo = pd.DataFrame(index=target.index)

grupo = valvula_ar_cleaner1.groupby(['gr']).mean()
for linha in grupo.index:
    modelo.loc[linha, grupo.columns[0]] = grupo.loc[linha, grupo.columns[0]]
valvula_ar_cleaner1_g = modelo

modelo = pd.DataFrame(index=target.index)

grupo = nivel_espuma_cleaner2.groupby(['gr']).mean()
for linha in grupo.index:
    modelo.loc[linha, grupo.columns[0]] = grupo.loc[linha, grupo.columns[0]]
nivel_espuma_cleaner2_g = modelo

modelo = pd.DataFrame(index=target.index)

grupo = vazao_ar_cleaner2.groupby(['gr']).mean()
for linha in grupo.index:
    modelo.loc[linha, grupo.columns[0]] = grupo.loc[linha, grupo.columns[0]]
vazao_ar_cleaner2_g = modelo

modelo = pd.DataFrame(index=target.index)

grupo = valvula_dardo1_cleaner2.groupby(['gr']).mean()
for linha in grupo.index:
    modelo.loc[linha, grupo.columns[0]] = grupo.loc[linha, grupo.columns[0]]
valvula_dardo1_cleaner2_g = modelo

modelo = pd.DataFrame(index=target.index)

grupo = valvula_dardo2_cleaner2.groupby(['gr']).mean()
for linha in grupo.index:
    modelo.loc[linha, grupo.columns[0]] = grupo.loc[linha, grupo.columns[0]]
valvula_dardo2_cleaner2_g = modelo

modelo = pd.DataFrame(index=target.index)

grupo = valvula_ar_cleaner2.groupby(['gr']).mean()
for linha in grupo.index:
    modelo.loc[linha, grupo.columns[0]] = grupo.loc[linha, grupo.columns[0]]
valvula_ar_cleaner2_g = modelo

modelo = pd.DataFrame(index=target.index)

grupo = nivel_espuma_recleaner.groupby(['gr']).mean()
for linha in grupo.index:
    modelo.loc[linha, grupo.columns[0]] = grupo.loc[linha, grupo.columns[0]]
nivel_espuma_recleaner_g = modelo

modelo = pd.DataFrame(index=target.index)

grupo = vazao_ar_recleaner.groupby(['gr']).mean()
for linha in grupo.index:
    modelo.loc[linha, grupo.columns[0]] = grupo.loc[linha, grupo.columns[0]]
vazao_ar_recleaner_g = modelo

modelo = pd.DataFrame(index=target.index)

grupo = valvula_dardo1_recleaner.groupby(['gr']).mean()
for linha in grupo.index:
    modelo.loc[linha, grupo.columns[0]] = grupo.loc[linha, grupo.columns[0]]
valvula_dardo1_recleaner_g = modelo

modelo = pd.DataFrame(index=target.index)

grupo = valvula_dardo2_recleaner.groupby(['gr']).mean()
for linha in grupo.index:
    modelo.loc[linha, grupo.columns[0]] = grupo.loc[linha, grupo.columns[0]]
valvula_dardo2_recleaner_g = modelo

modelo = pd.DataFrame(index=target.index)

grupo = valvula_ar_recleaner.groupby(['gr']).mean()
for linha in grupo.index:
    modelo.loc[linha, grupo.columns[0]] = grupo.loc[linha, grupo.columns[0]]
valvula_ar_recleaner_g = modelo

modelo = pd.DataFrame(index=target.index)

grupo = nivel_espuma_scavenger1b1.groupby(['gr']).mean()
for linha in grupo.index:
    modelo.loc[linha, grupo.columns[0]] = grupo.loc[linha, grupo.columns[0]]
nivel_espuma_scavenger1b1_g = modelo

modelo = pd.DataFrame(index=target.index)

grupo = vazao_ar_scavenger1b1.groupby(['gr']).mean()
for linha in grupo.index:
    modelo.loc[linha, grupo.columns[0]] = grupo.loc[linha, grupo.columns[0]]
vazao_ar_scavenger1b1_g = modelo

modelo = pd.DataFrame(index=target.index)

grupo = valvula_dardo1_scavenger1b1.groupby(['gr']).mean()
for linha in grupo.index:
    modelo.loc[linha, grupo.columns[0]] = grupo.loc[linha, grupo.columns[0]]
valvula_dardo1_scavenger1b1_g = modelo

modelo = pd.DataFrame(index=target.index)

grupo = valvula_dardo2_scavenger1b1.groupby(['gr']).mean()
for linha in grupo.index:
    modelo.loc[linha, grupo.columns[0]] = grupo.loc[linha, grupo.columns[0]]
valvula_dardo2_scavenger1b1_g = modelo

modelo = pd.DataFrame(index=target.index)

grupo = valvula_ar_scavenger1b1.groupby(['gr']).mean()
for linha in grupo.index:
    modelo.loc[linha, grupo.columns[0]] = grupo.loc[linha, grupo.columns[0]]
valvula_ar_scavenger1b1_g = modelo

modelo = pd.DataFrame(index=target.index)

grupo = nivel_espuma_scavenger1b2.groupby(['gr']).mean()
for linha in grupo.index:
    modelo.loc[linha, grupo.columns[0]] = grupo.loc[linha, grupo.columns[0]]
nivel_espuma_scavenger1b2_g = modelo

modelo = pd.DataFrame(index=target.index)

grupo = vazao_ar_scavenger1b2.groupby(['gr']).mean()
for linha in grupo.index:
    modelo.loc[linha, grupo.columns[0]] = grupo.loc[linha, grupo.columns[0]]
vazao_ar_scavenger1b2_g = modelo

modelo = pd.DataFrame(index=target.index)

grupo = valvula_dardo1_scavenger1b2.groupby(['gr']).mean()
for linha in grupo.index:
    modelo.loc[linha, grupo.columns[0]] = grupo.loc[linha, grupo.columns[0]]
valvula_dardo1_scavenger1b2_g = modelo

modelo = pd.DataFrame(index=target.index)

grupo = valvula_dardo2_scavenger1b2.groupby(['gr']).mean()
for linha in grupo.index:
    modelo.loc[linha, grupo.columns[0]] = grupo.loc[linha, grupo.columns[0]]
valvula_dardo2_scavenger1b2_g = modelo

modelo = pd.DataFrame(index=target.index)

grupo = valvula_ar_scavenger1b2.groupby(['gr']).mean()
for linha in grupo.index:
    modelo.loc[linha, grupo.columns[0]] = grupo.loc[linha, grupo.columns[0]]
valvula_ar_scavenger1b2_g = modelo

modelo = pd.DataFrame(index=target.index)

grupo = nivel_espuma_scavenger1b3.groupby(['gr']).mean()
for linha in grupo.index:
    modelo.loc[linha, grupo.columns[0]] = grupo.loc[linha, grupo.columns[0]]
nivel_espuma_scavenger1b3_g = modelo

modelo = pd.DataFrame(index=target.index)

grupo = vazao_ar_scavenger1b3.groupby(['gr']).mean()
for linha in grupo.index:
    modelo.loc[linha, grupo.columns[0]] = grupo.loc[linha, grupo.columns[0]]
vazao_ar_scavenger1b3_g = modelo

modelo = pd.DataFrame(index=target.index)

grupo = valvula_dardo1_scavenger1b3.groupby(['gr']).mean()
for linha in grupo.index:
    modelo.loc[linha, grupo.columns[0]] = grupo.loc[linha, grupo.columns[0]]
valvula_dardo1_scavenger1b3_g = modelo

modelo = pd.DataFrame(index=target.index)

grupo = valvula_dardo2_scavenger1b3.groupby(['gr']).mean()
for linha in grupo.index:
    modelo.loc[linha, grupo.columns[0]] = grupo.loc[linha, grupo.columns[0]]
valvula_dardo2_scavenger1b3_g = modelo

modelo = pd.DataFrame(index=target.index)

grupo = valvula_ar_scavenger1b3.groupby(['gr']).mean()
for linha in grupo.index:
    modelo.loc[linha, grupo.columns[0]] = grupo.loc[linha, grupo.columns[0]]
valvula_ar_scavenger1b3_g = modelo

modelo = pd.DataFrame(index=target.index)

grupo = nivel_espuma_scavenger2b1.groupby(['gr']).mean()
for linha in grupo.index:
    modelo.loc[linha, grupo.columns[0]] = grupo.loc[linha, grupo.columns[0]]
nivel_espuma_scavenger2b1_g = modelo

modelo = pd.DataFrame(index=target.index)

grupo = vazao_ar_scavenger2b1.groupby(['gr']).mean()
for linha in grupo.index:
    modelo.loc[linha, grupo.columns[0]] = grupo.loc[linha, grupo.columns[0]]
vazao_ar_scavenger2b1_g = modelo

modelo = pd.DataFrame(index=target.index)

grupo = valvula_dardo1_scavenger2b1.groupby(['gr']).mean()
for linha in grupo.index:
    modelo.loc[linha, grupo.columns[0]] = grupo.loc[linha, grupo.columns[0]]
valvula_dardo1_scavenger2b1_g = modelo

modelo = pd.DataFrame(index=target.index)

grupo = valvula_dardo2_scavenger2b1.groupby(['gr']).mean()
for linha in grupo.index:
    modelo.loc[linha, grupo.columns[0]] = grupo.loc[linha, grupo.columns[0]]
valvula_dardo2_scavenger2b1_g = modelo

modelo = pd.DataFrame(index=target.index)

grupo = valvula_ar_scavenger2b1.groupby(['gr']).mean()
for linha in grupo.index:
    modelo.loc[linha, grupo.columns[0]] = grupo.loc[linha, grupo.columns[0]]
valvula_ar_scavenger2b1_g = modelo

modelo = pd.DataFrame(index=target.index)

grupo = nivel_espuma_scavenger2b2.groupby(['gr']).mean()
for linha in grupo.index:
    modelo.loc[linha, grupo.columns[0]] = grupo.loc[linha, grupo.columns[0]]
nivel_espuma_scavenger2b2_g = modelo

modelo = pd.DataFrame(index=target.index)

grupo = vazao_ar_scavenger2b2.groupby(['gr']).mean()
for linha in grupo.index:
    modelo.loc[linha, grupo.columns[0]] = grupo.loc[linha, grupo.columns[0]]
vazao_ar_scavenger2b2_g = modelo

modelo = pd.DataFrame(index=target.index)

grupo = valvula_dardo1_scavenger2b2.groupby(['gr']).mean()
for linha in grupo.index:
    modelo.loc[linha, grupo.columns[0]] = grupo.loc[linha, grupo.columns[0]]
valvula_dardo1_scavenger2b2_g = modelo

modelo = pd.DataFrame(index=target.index)

grupo = valvula_dardo2_scavenger2b2.groupby(['gr']).mean()
for linha in grupo.index:
    modelo.loc[linha, grupo.columns[0]] = grupo.loc[linha, grupo.columns[0]]
valvula_dardo2_scavenger2b2_g = modelo

modelo = pd.DataFrame(index=target.index)

grupo = valvula_ar_scavenger2b2.groupby(['gr']).mean()
for linha in grupo.index:
    modelo.loc[linha, grupo.columns[0]] = grupo.loc[linha, grupo.columns[0]]
valvula_ar_scavenger2b2_g = modelo

modelo = pd.DataFrame(index=target.index)

grupo = nivel_espuma_scavenger2b3.groupby(['gr']).mean()
for linha in grupo.index:
    modelo.loc[linha, grupo.columns[0]] = grupo.loc[linha, grupo.columns[0]]
nivel_espuma_scavenger2b3_g = modelo

modelo = pd.DataFrame(index=target.index)

grupo = vazao_ar_scavenger2b3.groupby(['gr']).mean()
for linha in grupo.index:
    modelo.loc[linha, grupo.columns[0]] = grupo.loc[linha, grupo.columns[0]]
vazao_ar_scavenger2b3_g = modelo

modelo = pd.DataFrame(index=target.index)

grupo = valvula_dardo1_scavenger2b3.groupby(['gr']).mean()
for linha in grupo.index:
    modelo.loc[linha, grupo.columns[0]] = grupo.loc[linha, grupo.columns[0]]
valvula_dardo1_scavenger2b3_g = modelo

modelo = pd.DataFrame(index=target.index)

grupo = valvula_dardo2_scavenger2b3.groupby(['gr']).mean()
for linha in grupo.index:
    modelo.loc[linha, grupo.columns[0]] = grupo.loc[linha, grupo.columns[0]]
valvula_dardo2_scavenger2b3_g = modelo

modelo = pd.DataFrame(index=target.index)

grupo = valvula_ar_scavenger2b3.groupby(['gr']).mean()
for linha in grupo.index:
    modelo.loc[linha, grupo.columns[0]] = grupo.loc[linha, grupo.columns[0]]
valvula_ar_scavenger2b3_g = modelo

entrada = pd.concat([nivel_cx_alimentacao_g, densidade_cx_alimentacao_g], axis = 1)
entrada = pd.concat([entrada, vazao_vol_cx_alimentacao_g], axis = 1)
entrada = pd.concat([entrada, vazao_mas_cx_alimentacao_g], axis = 1)
entrada = pd.concat([entrada, vazao_amina_cx_alimentacao_g], axis = 1)
entrada = pd.concat([entrada, vazao_amido_cx_alimentacao_g], axis = 1)
entrada = pd.concat([entrada, nivel_cx_recirculacao_g], axis = 1)
entrada = pd.concat([entrada, densidade_cx_recirculacao_g], axis = 1)
entrada = pd.concat([entrada, vazao_vol_cx_recirculacao_g], axis = 1)
entrada = pd.concat([entrada, vazao_mas_cx_recirculacao_g], axis = 1)
entrada = pd.concat([entrada, nivel_espuma_rougher1_g], axis = 1)
entrada = pd.concat([entrada, vazao_ar_rougher1_g], axis = 1)
entrada = pd.concat([entrada, valvula_dardo1_rougher1_g], axis = 1)
entrada = pd.concat([entrada, valvula_dardo2_rougher1_g], axis = 1)
entrada = pd.concat([entrada, valvula_ar_rougher1_g], axis = 1)
entrada = pd.concat([entrada, nivel_espuma_rougher2_g], axis = 1)
entrada = pd.concat([entrada, vazao_ar_rougher2_g], axis = 1)
entrada = pd.concat([entrada, valvula_dardo1_rougher2_g], axis = 1)
entrada = pd.concat([entrada, valvula_dardo2_rougher2_g], axis = 1)
entrada = pd.concat([entrada, valvula_ar_rougher2_g], axis = 1)
entrada = pd.concat([entrada, nivel_espuma_cleaner1_g], axis = 1)
entrada = pd.concat([entrada, vazao_ar_cleaner1_g], axis = 1)
entrada = pd.concat([entrada, valvula_dardo1_cleaner1_g], axis = 1)
entrada = pd.concat([entrada, valvula_dardo2_cleaner1_g], axis = 1)
entrada = pd.concat([entrada, valvula_ar_cleaner1_g], axis = 1)
entrada = pd.concat([entrada, nivel_espuma_cleaner2_g], axis = 1)
entrada = pd.concat([entrada, vazao_ar_cleaner2_g], axis = 1)
entrada = pd.concat([entrada, valvula_dardo1_cleaner2_g], axis = 1)
entrada = pd.concat([entrada, valvula_dardo2_cleaner2_g], axis = 1)
entrada = pd.concat([entrada, valvula_ar_cleaner2_g], axis = 1)
entrada = pd.concat([entrada, nivel_espuma_recleaner_g], axis = 1)
entrada = pd.concat([entrada, vazao_ar_recleaner_g], axis = 1)
entrada = pd.concat([entrada, valvula_dardo1_recleaner_g], axis = 1)
entrada = pd.concat([entrada, valvula_dardo2_recleaner_g], axis = 1)
entrada = pd.concat([entrada, valvula_ar_recleaner_g], axis = 1)
entrada = pd.concat([entrada, nivel_espuma_scavenger1b1_g], axis = 1)
entrada = pd.concat([entrada, vazao_ar_scavenger1b1_g], axis = 1)
entrada = pd.concat([entrada, valvula_dardo1_scavenger1b1_g], axis = 1)
entrada = pd.concat([entrada, valvula_dardo2_scavenger1b1_g], axis = 1)
entrada = pd.concat([entrada, valvula_ar_scavenger1b1_g], axis = 1)
entrada = pd.concat([entrada, nivel_espuma_scavenger1b2_g], axis = 1)
entrada = pd.concat([entrada, vazao_ar_scavenger1b2_g], axis = 1)
entrada = pd.concat([entrada, valvula_dardo1_scavenger1b2_g], axis = 1)
entrada = pd.concat([entrada, valvula_dardo2_scavenger1b2_g], axis = 1)
entrada = pd.concat([entrada, valvula_ar_scavenger1b2_g], axis = 1)
entrada = pd.concat([entrada, nivel_espuma_scavenger1b3_g], axis = 1)
entrada = pd.concat([entrada, vazao_ar_scavenger1b3_g], axis = 1)
entrada = pd.concat([entrada, valvula_dardo1_scavenger1b3_g], axis = 1)
entrada = pd.concat([entrada, valvula_dardo2_scavenger1b3_g], axis = 1)
entrada = pd.concat([entrada, valvula_ar_scavenger1b3_g], axis = 1)
#entrada = pd.concat([entrada, nivel_espuma_scavenger2b1_g], axis = 1)
#entrada = pd.concat([entrada, vazao_ar_scavenger2b1_g], axis = 1)
#entrada = pd.concat([entrada, valvula_dardo1_scavenger2b1_g], axis = 1)
#entrada = pd.concat([entrada, valvula_dardo2_scavenger2b1_g], axis = 1)
#entrada = pd.concat([entrada, valvula_ar_scavenger2b1_g], axis = 1)
entrada = pd.concat([entrada, nivel_espuma_scavenger2b2_g], axis = 1)
entrada = pd.concat([entrada, vazao_ar_scavenger2b2_g], axis = 1)
entrada = pd.concat([entrada, valvula_dardo1_scavenger2b2_g], axis = 1)
entrada = pd.concat([entrada, valvula_dardo2_scavenger2b2_g], axis = 1)
entrada = pd.concat([entrada, valvula_ar_scavenger2b2_g], axis = 1)
entrada = pd.concat([entrada, nivel_espuma_scavenger2b3_g], axis = 1)
entrada = pd.concat([entrada, vazao_ar_scavenger2b3_g], axis = 1)
entrada = pd.concat([entrada, valvula_dardo1_scavenger2b3_g], axis = 1)
entrada = pd.concat([entrada, valvula_dardo2_scavenger2b3_g], axis = 1)
entrada = pd.concat([entrada, valvula_ar_scavenger2b3_g], axis = 1)

entrada.drop(index=-1, inplace=True)

agrupados.shape

agrupados.to_csv('agrupados.csv', ',', index=False)

entrada.iloc[0:5,59:70]

agrupados.iloc[0:5,45:50]
