==> 04b-Agrupamentos-Otimizado.ipynb <==
# Importa as bibliotecas
import pandas as pd
import numpy as np

pasta = 'Delay/05-Maio/'

# Faz a leitura dos dados
target = pd.read_excel(pasta + 'T_Conceicao.xlsx')


nivel_cx_alimentacao = pd.read_excel(pasta + 'CA_Nivel.xlsx')

densidade_cx_alimentacao = pd.read_excel(pasta + 'CA_Densidade.xlsx')

vazao_vol_cx_alimentacao = pd.read_excel(pasta + 'CA_Vazao_Vol.xlsx')

# vazao_mas_cx_alimentacao = pd.read_excel(pasta + 'CA_Vazao_Mas.xlsx')

vazao_amina_cx_alimentacao = pd.read_excel(pasta + 'CA_Vazao_Amina.xlsx')

vazao_amido_cx_alimentacao = pd.read_excel(pasta + 'CA_Vazao_Amido.xlsx')

# lista = [nivel_cx_alimentacao, densidade_cx_alimentacao, vazao_vol_cx_alimentacao,
#         vazao_mas_cx_alimentacao, vazao_amina_cx_alimentacao, vazao_amido_cx_alimentacao]
lista = [nivel_cx_alimentacao, densidade_cx_alimentacao, vazao_vol_cx_alimentacao,
         vazao_amina_cx_alimentacao, vazao_amido_cx_alimentacao]

# nivel_cx_recirculacao = pd.read_excel(pasta + 'CR_Nivel.xlsx')

densidade_cx_recirculacao = pd.read_excel(pasta + 'CR_Densidade.xlsx')

vazao_vol_cx_recirculacao = pd.read_excel(pasta + 'CR_Vazao_Vol.xlsx')

# vazao_mas_cx_recirculacao = pd.read_excel(pasta + 'CR_Vazao_Mas.xlsx')

# lista.append(nivel_cx_recirculacao)
lista.append(densidade_cx_recirculacao)
lista.append(vazao_vol_cx_recirculacao)
# lista.append(vazao_mas_cx_recirculacao)

nivel_espuma_rougher1 = pd.read_excel(pasta + 'R1_Nivel_Espuma.xlsx')

# vazao_ar_rougher1 = pd.read_excel(pasta + 'R1_Vazao_Ar.xlsx')

# valvula_dardo1_rougher1 = pd.read_excel(pasta + 'R1_Valv_Dardo1.xlsx')

# valvula_dardo2_rougher1 = pd.read_excel(pasta + 'R1_Valv_Dardo2.xlsx')

# valvula_ar_rougher1 = pd.read_excel(pasta + 'R1_Valv_Ar.xlsx')

#Verificar cabeçalhos
lista.append(nivel_espuma_rougher1)
# lista.append(vazao_ar_rougher1)
# lista.append(valvula_dardo1_rougher1)
# lista.append(valvula_dardo2_rougher1)
# lista.append(valvula_ar_rougher1)

nivel_espuma_rougher2 = pd.read_excel(pasta + 'R2_Nivel_Espuma.xlsx')

# vazao_ar_rougher2 = pd.read_excel(pasta + 'R2_Vazao_Ar.xlsx')

# valvula_dardo1_rougher2 = pd.read_excel(pasta + 'R2_Valv_Dardo1.xlsx')

# valvula_dardo2_rougher2 = pd.read_excel(pasta + 'R2_Valv_Dardo2.xlsx')

# valvula_ar_rougher2 = pd.read_excel(pasta + 'R2_Valv_Ar.xlsx')

lista.append(nivel_espuma_rougher2)
# lista.append(vazao_ar_rougher2)
# lista.append(valvula_dardo1_rougher2)
# lista.append(valvula_dardo2_rougher2)
# lista.append(valvula_ar_rougher2)

nivel_espuma_cleaner1 = pd.read_excel(pasta + 'C1_Nivel_Espuma.xlsx')

# vazao_ar_cleaner1 = pd.read_excel(pasta + 'C1_Vazao_Ar.xlsx')

valvula_dardo1_cleaner1 = pd.read_excel(pasta + 'C1_Valv_Dardo1.xlsx')

# valvula_dardo2_cleaner1 = pd.read_excel(pasta + 'C1_Valv_Dardo2.xlsx')

# valvula_ar_cleaner1 = pd.read_excel(pasta + 'C1_Valv_Ar.xlsx')

lista.append(nivel_espuma_cleaner1)
# lista.append(vazao_ar_cleaner1)
lista.append(valvula_dardo1_cleaner1)
# lista.append(valvula_dardo2_cleaner1)
# lista.append(valvula_ar_cleaner1)

# nivel_espuma_cleaner2 = pd.read_excel(pasta + 'C2_Nivel_Espuma.xlsx')

# vazao_ar_cleaner2 = pd.read_excel(pasta + 'C2_Vazao_Ar.xlsx')

# valvula_dardo1_cleaner2 = pd.read_excel(pasta + 'C2_Valv_Dardo1.xlsx')

# valvula_dardo2_cleaner2 = pd.read_excel(pasta + 'C2_Valv_Dardo2.xlsx')

# valvula_ar_cleaner2 = pd.read_excel(pasta + 'C2_Valv_Ar.xlsx')

# lista.append(nivel_espuma_cleaner2)
# lista.append(vazao_ar_cleaner2)
# lista.append(valvula_dardo1_cleaner2)
# lista.append(valvula_dardo2_cleaner2)
# lista.append(valvula_ar_cleaner2)

nivel_espuma_recleaner = pd.read_excel(pasta + 'RC_Nivel_Espuma.xlsx')

# vazao_ar_recleaner = pd.read_excel(pasta + 'RC_Vazao_Ar.xlsx')

# valvula_dardo1_recleaner = pd.read_excel(pasta + 'RC_Valv_Dardo1.xlsx')

# valvula_dardo2_recleaner = pd.read_excel(pasta + 'RC_Valv_Dardo2.xlsx')

# valvula_ar_recleaner = pd.read_excel(pasta + 'RC_Valv_Ar.xlsx')

lista.append(nivel_espuma_recleaner)
# lista.append(vazao_ar_recleaner)
# lista.append(valvula_dardo1_recleaner)
# lista.append(valvula_dardo2_recleaner)
# lista.append(valvula_ar_recleaner)

# nivel_espuma_scavenger1b1 = pd.read_excel(pasta + 'S1B1_Nivel_Espuma.xlsx')

# vazao_ar_scavenger1b1 = pd.read_excel(pasta + 'S1B1_Vazao_Ar.xlsx') TUDO ZERO

# valvula_dardo1_scavenger1b1 = pd.read_excel(pasta + 'S1B1_Valv_Dardo1.xlsx')

# valvula_dardo2_scavenger1b1 = pd.read_excel(pasta + 'S1B1_Valv_Dardo2.xlsx')

# valvula_ar_scavenger1b1 = pd.read_excel(pasta + 'S1B1_Valv_Ar.xlsx')

# lista.append(nivel_espuma_scavenger1b1)
# lista.append(vazao_ar_scavenger1b1) TUDO ZERO
# lista.append(valvula_dardo1_scavenger1b1)
# lista.append(valvula_dardo2_scavenger1b1)
# lista.append(valvula_ar_scavenger1b1)

nivel_espuma_scavenger1b2 = pd.read_excel(pasta + 'S1B2_Nivel_Espuma.xlsx')

# vazao_ar_scavenger1b2 = pd.read_excel(pasta + 'S1B2_Vazao_Ar.xlsx')

# valvula_dardo1_scavenger1b2 = pd.read_excel(pasta + 'S1B2_Valv_Dardo1.xlsx')

# valvula_dardo2_scavenger1b2 = pd.read_excel(pasta + 'S1B2_Valv_Dardo2.xlsx')

# valvula_ar_scavenger1b2 = pd.read_excel(pasta + 'S1B2_Valv_Ar.xlsx')

lista.append(nivel_espuma_scavenger1b2)
# lista.append(vazao_ar_scavenger1b2)
# lista.append(valvula_dardo1_scavenger1b2)
# lista.append(valvula_dardo2_scavenger1b2)
# lista.append(valvula_ar_scavenger1b2)

nivel_espuma_scavenger1b3 = pd.read_excel(pasta + 'S1B3_Nivel_Espuma.xlsx')

# vazao_ar_scavenger1b3 = pd.read_excel(pasta + 'S1B3_Vazao_Ar.xlsx') TUDO ZERO

# valvula_dardo1_scavenger1b3 = pd.read_excel(pasta + 'S1B3_Valv_Dardo1.xlsx')

# valvula_dardo2_scavenger1b3 = pd.read_excel(pasta + 'S1B3_Valv_Dardo2.xlsx')

# valvula_ar_scavenger1b3 = pd.read_excel(pasta + 'S1B3_Valv_Ar.xlsx')

lista.append(nivel_espuma_scavenger1b3)
# lista.append(vazao_ar_scavenger1b3) TUDO ZERO
# lista.append(valvula_dardo1_scavenger1b3)
# lista.append(valvula_dardo2_scavenger1b3)
# lista.append(valvula_ar_scavenger1b3)

# nivel_espuma_scavenger2b1 = pd.read_excel(pasta + 'S2B1_Nivel_Espuma.xlsx')

# vazao_ar_scavenger2b1 = pd.read_excel(pasta + 'S2B1_Vazao_Ar.xlsx') TUDO ZERO

# valvula_dardo1_scavenger2b1 = pd.read_excel(pasta + 'S2B1_Valv_Dardo1.xlsx')

# valvula_dardo2_scavenger2b1 = pd.read_excel(pasta + 'S2B1_Valv_Dardo2.xlsx')

# valvula_ar_scavenger2b1 = pd.read_excel(pasta + 'S2B1_Valv_Ar.xlsx')

# lista.append(nivel_espuma_scavenger2b1)
# lista.append(vazao_ar_scavenger2b1) TUDO ZERO
# lista.append(valvula_dardo1_scavenger2b1)
# lista.append(valvula_dardo2_scavenger2b1)
# lista.append(valvula_ar_scavenger2b1)

# nivel_espuma_scavenger2b2 = pd.read_excel(pasta + 'S2B2_Nivel_Espuma.xlsx')

# vazao_ar_scavenger2b2 = pd.read_excel(pasta + 'S2B2_Vazao_Ar.xlsx')

# valvula_dardo1_scavenger2b2 = pd.read_excel(pasta + 'S2B2_Valv_Dardo1.xlsx')

# valvula_dardo2_scavenger2b2 = pd.read_excel(pasta + 'S2B2_Valv_Dardo2.xlsx')

# valvula_ar_scavenger2b2 = pd.read_excel(pasta + 'S2B2_Valv_Ar.xlsx')

# lista.append(nivel_espuma_scavenger2b2)
# lista.append(vazao_ar_scavenger2b2)
# lista.append(valvula_dardo1_scavenger2b2)
# lista.append(valvula_dardo2_scavenger2b2)
# lista.append(valvula_ar_scavenger2b2)

# nivel_espuma_scavenger2b3 = pd.read_excel(pasta + 'S2B3_Nivel_Espuma.xlsx')

vazao_ar_scavenger2b3 = pd.read_excel(pasta + 'S2B3_Vazao_Ar.xlsx')

# valvula_dardo1_scavenger2b3 = pd.read_excel(pasta + 'S2B3_Valv_Dardo1.xlsx')

# valvula_dardo2_scavenger2b3 = pd.read_excel(pasta + 'S2B3_Valv_Dardo2.xlsx')

# valvula_ar_scavenger2b3 = pd.read_excel(pasta + 'S2B3_Valv_Ar.xlsx')

# lista.append(nivel_espuma_scavenger2b3)
lista.append(vazao_ar_scavenger2b3)
# lista.append(valvula_dardo1_scavenger2b3)
# lista.append(valvula_dardo2_scavenger2b3)
# lista.append(valvula_ar_scavenger2b3)

agrupados = pd.DataFrame(index=target.index)
preenchidos = pd.DataFrame(index=target.index)
for item in lista:
    modelo = pd.DataFrame(index=target.index)
    grupo = item.groupby(['gr']).median()
    for linha in grupo.index:
        modelo.loc[linha, grupo.columns[0]] = grupo.loc[linha, grupo.columns[0]]
    agrupados = pd.concat([agrupados, modelo], axis = 1)
    
    repetir = False
    anterior = 0
    proximo = -1
    # Determina quais linhas são NA
    valores_na = modelo.isna()
    for linha in target.index:
        try:
            # Se a linha for NA
            if valores_na.iloc[linha].bool():
                # Se não estiver marcado para repetir o valor anterior
                if not repetir:
                    # Se não está na última linha
                    if linha < target.index.shape[0]:
                        # Busca o próximo valor na coluna
                        for busca_valor in range(linha+1, target.index.shape[0]):
                            # Se é encontrado um valor (não-NA)
                            if not valores_na.iloc[busca_valor].bool():
                                # Esse valor é definido como o 'próximo'
                                proximo = busca_valor
                                break
                    # Se existe um valor anterior à linha atual
                    if anterior != linha:
                        # Se existe um valor seguinte à linha atual (Busca próximo valor encontrou um resultado)
                        if proximo != -1:
                            # Define o valor atual como a média entre o anterior e o próximo
                            valor = np.mean([modelo.iloc[anterior,0], modelo.iloc[proximo,0]])
                        # Se não existe um valor seguinte à linha atual (Busca próximo valor não encontrou um resultado)
                        else:
                            # Repete o valor anterior
                            valor = modelo.iloc[anterior,0]
                    # Se não existe um valor anterior à linha atual
                    else:
                        # Se existe um valor seguinte à linha atual (Busca próximo valor encontrou um resultado)
                        if proximo != -1:
                            valor = modelo.iloc[proximo,0]
                        # Se não existe um valor seguinte à linha atual (Busca próximo valor não encontrou um resultado)
                        else:
                            print('Problema na linha', linha)

                # Atribui o valor à linha atual
                modelo.iloc[linha,0] = valor

                # Se o próximo valor não estiver na linha seguinte
                if proximo > linha + 1:
                    # Marca para repetir o valor atual
                    repetir = True
                # Se o próximo valor estiver na linha seguinte
                else:
                    repetir = False
                    proximo = -1

            anterior = linha
        except:
            pass
        
    
    preenchidos = pd.concat([preenchidos, modelo], axis = 1)
try:
    agrupados.drop(index=-1, inplace=True)
    preenchidos.drop(index=-1, inplace=True)
except:
    pass

agrupados.to_excel(pasta + '04-ag_mediana_red_interd.xlsx', 'Planilha1', index=False)
preenchidos.to_excel(pasta + '04-ag_preenchidos_mediana_red_interd.xlsx', 'Planilha1', index=False)

#agrupados.to_excel(pasta + '04-agrupados_mediana_red_interd.xlsx', 'Planilha1', index=False)
#preenchidos.to_excel(pasta + '04-ag_preenchidos_mediana_red_interd.xlsx', 'Planilha1', index=False)

#agrupados.to_excel(pasta + 'agrupados_mediana_red.xlsx', 'Planilha1', index=False)
#preenchidos.to_excel(pasta + 'ag_preenchidos_mediana_red.xlsx', 'Planilha1', index=False)

#agrupados.to_excel(pasta + 'agrupados_media.xlsx', 'Planilha1', index=False)
#preenchidos.to_excel(pasta + 'ag_preenchidos_media.xlsx', 'Planilha1', index=False)
