==> 07d-Normalizacao-Melhorada_v1.0.ipynb <==
from __future__ import absolute_import, division, print_function, unicode_literals

import pathlib

import matplotlib.axes as mtpaxes
import matplotlib.pyplot as plt
import pandas as pd
import seaborn as sns

# Gráficos do matplotlib interativos
%matplotlib notebook

### Get the data

pasta = 'Experimentos/03-Marco/'
#pasta = 'Delay/05-Maio/'
raw_input = pd.read_excel(pasta + '04-ag_preenchidos_mediana_red_interd_15.xlsx', header=0)
#raw_input = pd.read_excel(pasta + '04-ag_preenchidos_mediana_red_interd_15.xlsx', header=0)
raw_target = pd.read_excel(pasta + 'T_Conceicao.xlsx', header=0)

dataset = pd.concat([raw_input, raw_target.iloc[:,1]], axis=1)
#dataset_mar

### Get the reference data

pasta_mar = 'Delay/03-Marco/'
#raw_input_mar = pd.read_excel(pasta_mar + '04-ag_preenchidos_mediana_red_interd.xlsx', header=0)
#raw_target_mar = pd.read_excel(pasta_mar + 'T_Conceicao.xlsx', header=0)

dataset_mar = pd.concat([raw_input_mar, raw_target_mar.iloc[:,1]], axis=1)
#dataset_mar

#train_stats = train_dataset.describe()
train_stats = dataset_mar.describe()
train_stats.pop('Ferro')
train_stats = train_stats.transpose()
train_stats

raw_input_mar = pd.read_excel(pasta_mar + '00-Faixas.xlsx', header=0)
raw_input_mar = raw_input_mar.transpose()
# Renomeia o cabeçalho de target
raw_input_mar.columns = ['max', 'min']

train_stats = raw_input_mar

### Normalize the data

Look again at the `train_stats` block above and note how different the ranges of each feature are.

normed_dataset = pd.DataFrame(index=dataset.index)
for parametro in dataset.drop(columns='Ferro').columns:
    param = (dataset[parametro] - train_stats['min'].loc[parametro]) / (train_stats['max'].loc[parametro] - 
                                                                        train_stats['min'].loc[parametro])
    
    normed_dataset = pd.concat([normed_dataset, param], axis = 1)

normed_dataset.fillna(0,inplace=True)

## Conclusion

normed_dataset.to_excel(pasta + '05-mediana_red_final_norm.xlsx', 'Planilha1', index=False)

normed_dataset.to_excel(pasta + '06-mediana_red_final_norm_independente.xlsx', 'Planilha1', index=False)

normed_dataset.to_excel(pasta + '06-mediana_red_final_norm_extrapolado.xlsx', 'Planilha1', index=False)
